% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{alphacolor}
\alias{alphacolor}
\title{alphacolor}
\usage{
alphacolor(cols, a = 0.99)
}
\arguments{
\item{cols}{Colors vector}

\item{a}{Alpha level numeric ranging from 0 to 1}
}
\description{
Returns colors at alpha level a
}

