% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-memberships.R
\name{as_screen_name}
\alias{as_screen_name}
\title{as screen name}
\usage{
as_screen_name(x)
}
\arguments{
\item{x}{Vector of users (screen names).}
}
\value{
Vector of class screen_name.
}
\description{
Forces user of type screen name.
}
