% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_id.R
\name{as_screenname}
\alias{as_screenname}
\alias{as_userid}
\title{Mark a user id as a screen name}
\usage{
as_screenname(x)
}
\arguments{
\item{x}{A character vector of twitter screen names.}
}
\description{
There are two ways to identify a twitter user: a screen name (e.g.
"justinbieber") or a user identifier (e.g. "27260086"). User identifiers
are 64-bit integers which can't be represented natively by any base R data
structures, so in most cases rtweet returns them as strings. This introduces
an ambiguity, because user names can also consist solely of numbers
(e.g. "123456"). You can resolve this ambiguity by using \code{as_screenname()}
to a character vector a screen name.
}
\examples{
\dontrun{
# Look up user with id 123456 (screen name harperreed)
lookup_users("123456") 

# Look up user with name 123456
lookup_users(as_screenname("123456"))
}
}
\seealso{
Other users: 
\code{\link{lists_subscribers}()},
\code{\link{lookup_users}()},
\code{\link{search_users}()},
\code{\link{tweets_with_users}()},
\code{\link{users_data}()}
}
\concept{users}
