% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-memberships.R
\name{as_user_id}
\alias{as_user_id}
\title{as user ID}
\usage{
as_user_id(x)
}
\arguments{
\item{x}{Vector of users (user IDs).}
}
\value{
Vector of class user_Id.
}
\description{
Forces user of type user ID.
}
