% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_as}
\alias{auth_as}
\title{Set default authentication}
\usage{
auth_as(auth = NULL)
}
\arguments{
\item{auth}{One of the following options:
\itemize{
\item \code{NULL}, the default, will look for cached authentication from a previous
session. If none are found, will automatically set up \code{\link[=rtweet_user]{rtweet_user()}};
if one is found, will use that; if multiple are found, it will error.
\item A string giving the name of a cached auth file.
\item An object created by \code{\link[=rtweet_app]{rtweet_app()}}, \code{\link[=rtweet_bot]{rtweet_bot()}}, or \code{\link[=rtweet_user]{rtweet_user()}}.
}}
}
\value{
Invisibly returns the previous authentication mechanism.
}
\description{
\code{auth_as()} sets up the default authentication mechanism used by all
rtweet API calls. See \code{\link[=rtweet_user]{rtweet_user()}} to learn more about the three
available authentication options.
}
\examples{
\dontrun{
# Use app auth for the remainder of this session:
my_app <- rtweet_app()
auth_as(my_app)

# Switch back to the default user based auth
auth_as()

# Load auth saved by auth_save()
auth_as("my-saved-app")
}
}
