% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_clean.R
\name{auth_clean}
\alias{auth_clean}
\title{Remove tokens}
\usage{
auth_clean(old = TRUE, new = FALSE)
}
\arguments{
\item{old}{A logical value if you want to remove old tokens.}

\item{new}{A logical value if you want to remove new tokens.}
}
\value{
An invisible logical value showing the success of the operation.
If no tokens need to be deleted it will return FALSE too.
\code{NULL} if there is nothing to do.
}
\description{
If there is a file with saved tokens it will delete it.
}
\details{
This functions helps to comply with CRAN policy to remove files created by
the package.
}
\examples{
auth_clean(TRUE, TRUE)
}
