% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_get}
\alias{auth_get}
\title{Get the current authentication mechanism}
\usage{
auth_get()
}
\description{
If no authentcation has been set up for this session, will call \code{\link[=auth_as]{auth_as()}}
to set up.
}
\keyword{internal}
