% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_save}
\alias{auth_save}
\title{Save an authentication mechanism for use in a future session}
\usage{
auth_save(auth, name)
}
\arguments{
\item{auth}{One of \code{\link[=rtweet_app]{rtweet_app()}}, \code{\link[=rtweet_bot]{rtweet_bot()}}, or \code{\link[=rtweet_user]{rtweet_user()}}.}

\item{name}{Cache name to use.}
}
\description{
Use \code{auth_save()} with \code{\link[=auth_as]{auth_as()}} to avoid repeatedly entering app
credentials, making it easier to share auth between projects.
}
\examples{
\dontrun{
# save app auth for use in other sessions
auth <- rtweet_app()
auth_save(auth, "my-app")

# later, in a different session...
auth_us("my-app")
}
}
