% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_setup_default}
\alias{auth_setup_default}
\alias{auth_has_default}
\title{Set up default authentication}
\usage{
auth_setup_default()

auth_has_default()
}
\description{
You'll need to run this function once per computer so that rtweet can use
your personal twitter account. See \code{\link[=rtweet_app]{rtweet_app()}}/\link{rtweet_bot} and
\code{\link[=auth_save]{auth_save()}} for other authentication options.
}
\seealso{
Other authentication: 
\code{\link{auth_as}()},
\code{\link{auth_get}()},
\code{\link{auth_save}()},
\code{\link{rtweet_user}()}
}
\concept{authentication}
