% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_sitrep.R
\name{auth_sitrep}
\alias{auth_sitrep}
\title{Twitter Tokens sitrep}
\usage{
auth_sitrep()
}
\value{
Invisibly, TRUE if some problems were found and FALSE otherwise
}
\description{
Get a situation report of your current tokens; useful for upgrading from
rtweet 0.7.0 to 1.0.0 and diagnosing problems with tokens.
}
\examples{
auth_sitrep()
}
\seealso{
\code{\link[=auth_as]{auth_as()}}
}
