% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_user}
\alias{auth_user}
\alias{auth_bot}
\alias{auth_app}
\title{Authentication options}
\usage{
auth_user(api_key = NULL, api_secret = NULL)

auth_bot(api_key, api_secret, access_token, access_secret)

auth_app(bearer_token)
}
\arguments{
\item{api_key, api_secret}{App API key and secret. These are generally not
required for \code{auth_user()} since if not supplied it will use the built-in
rtweet app.}

\item{access_token, access_secret}{Access token and secret. Together with
\code{api_key} and \code{api_secret} these give full control over your app, so
should not be stored in source code. Instead store them in environment
variables and retrieve with \code{\link[=Sys.getenv]{Sys.getenv()}}}

\item{bearer_token}{App bearer token.}
}
\description{
There are three ways that you can authenticate with the twitter API:
\itemize{
\item \code{auth_user()} interactively authenticates an existing twitter user.
This form is most appropriate if you want rtweet to control your
twitter account.
\item \code{auth_app()} authenticates as a twitter application. An application can't
perform actions (i.e. it can't tweet) but otherwise has generally higher
rate limits (i.e. you can do more searches) generally higher. This form is
most appropriate if you are collecting data.
\item \code{auth_bot()} authenticates as bot that takes actions on behalf of an app.
This form is most appropriate if you want to create a twitter account that
is run by a computer, rather than a human.
}

To use \code{auth_app()} or \code{auth_bot()} you will need to create your own Twitter
application following the instructions in \code{vignette("auth.Rmd")}.

See \code{\link[=use_auth]{use_auth()}} to set an auth mechanism as default for this session, and
\code{\link[=save_auth]{save_auth()}} to save an authentication mechanism so you can use it across
multiple sessions.
}
