% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{basemap}
\alias{basemap}
\title{basemap}
\usage{
basemap(mapdf = NULL, xlim = c(-179, 179), ylim = c(-55, 80),
  ocean.col = "#e0e0eb99", land.col = "#ffffff", border.lwd = 0.1)
}
\arguments{
\item{mapdf}{Mapping data frame defaults to high resolution
rworldmap.}

\item{xlim}{Maximimum plot values along x-axis. Defaults to
\code{xlim = c(-145, 145)} with a max range of -180 to
180 (degree longitude)}

\item{ylim}{Maximimum plot values along y-axis. Defaults to
\code{ylim = c(-50, 70)} with a max range of -90 to
90 (degrees latitude)}

\item{ocean.col}{Fill color to apply to ocean area to a
light gray color.}

\item{land.col}{Fill color to apply to land area defaults to
white.}

\item{border.lwd}{Width of country borders defaults to .025}
}
\value{
Plot of world map.
}
\description{
Plots world map
}

