% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-block.R
\name{block_user}
\alias{block_user}
\alias{unblock_user}
\title{Blocking or unblocking twitter users}
\usage{
block_user(user, unblock = FALSE, token = NULL)
}
\arguments{
\item{user}{Character vector of screen names or user ids.
See \code{\link[=as_screenname]{as_screenname()}} for more details.}

\item{unblock}{Logical indicating whether to unblock the intended
friend}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\description{
\code{block_user(...)} blocks or unblocks a target twitter user.
\code{unblock_user(...)} is synonymous to \code{block_user(..., unblock=TRUE)}
}
\examples{
\dontrun{
block_user("BarackObama")
unblock_user("BarackObama")
block_user("BarackObama", unblock=TRUE) #<-same as the above
}
}
\references{
Block: \url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/mute-block-report-users/api-reference/post-blocks-create}
}
\seealso{
Other post: 
\code{\link{post_favorite}()},
\code{\link{post_follow}()},
\code{\link{post_friendship}()},
\code{\link{post_tweet}()}
}
\concept{post}
