% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{check_chunked_media_status}
\alias{check_chunked_media_status}
\title{Checks status of chunked media upload`}
\usage{
check_chunked_media_status(finalize_data, token, rurl)
}
\arguments{
\item{finalize_data}{Output of FINALIZE or STATUS command.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}

\item{rurl}{Upload address for media.}
}
\description{
Checks status of chunked media upload`
}
