% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{check_token}
\alias{check_token}
\title{check_token}
\usage{
check_token(token, query = NULL)
}
\arguments{
\item{token}{user token}

\item{query}{user of interest}
}
\description{
check_token
}
\keyword{internal}

