% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woeids.R
\name{check_woeid}
\alias{check_woeid}
\title{check_woeid}
\usage{
check_woeid(x)
}
\arguments{
\item{x}{Character string, place}
}
\description{
Returns WOEID number for desired town or country.
}
\keyword{internal}

