% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_clean}
\alias{client_clean}
\title{Clean clients}
\usage{
client_clean()
}
\value{
Either TRUE if folder could be deleted or FALSE.
}
\description{
Removes the rtweet clients directory.
}
\examples{
client_clean()
}
