% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{client_get}
\alias{client_get}
\title{Get the current client}
\usage{
client_get()
}
\value{
The current client used.
}
\description{
If no client has been set up for this session, \code{client_get()} will
call \code{\link[=client_as]{client_as()}} to set it up.
}
\examples{
\dontrun{
client_get()
}
}
\seealso{
Other client: 
\code{\link{client_as}()},
\code{\link{client_has_default}()},
\code{\link{client_save}()}
}
\concept{client}
