% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_messages.R
\name{direct_messages}
\alias{direct_messages}
\alias{direct_messages_sent}
\title{Get direct messages sent to and received by the authenticating user from the
past 30 days}
\usage{
direct_messages(
  n = 50,
  cursor = NULL,
  next_cursor = NULL,
  parse = TRUE,
  token = NULL
)

direct_messages_sent(
  since_id = NULL,
  max_id = NULL,
  n = 200,
  parse = TRUE,
  token = NULL
)
}
\arguments{
\item{n}{Maximum number of results to return.}

\item{cursor}{Which page of results to return. The default will return
the first page; can be used for manual pagination.}

\item{next_cursor}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{cursor} instead.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{since_id}{optional Returns results with an ID greater than
(that is, more recent than) the specified ID. There are limits to
the number of Tweets which can be accessed through the API. If
the limit of Tweets has occurred since the since_id, the since_id
will be forced to the oldest ID available.}

\item{max_id}{String giving id of most recent tweet to return.
Can be used for manual pagination.}
}
\value{
A list with one element for each page of results.
}
\description{
Returns all Direct Message events (both sent and received) within the last 30
days. Sorted in reverse-chronological order. Includes detailed information
about the sender and recipient.
}
\examples{

\dontrun{

## get my direct messages
dms <- direct_messages()

## inspect data structure
str(dms)

}
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/direct-messages/sending-and-receiving/api-reference/list-events}
}
