% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emojis}
\alias{emojis}
\title{Defunct: Emojis codes and descriptions data.}
\format{
A tibble with two variables and 2,623 observations.
}
\description{
This data comes from "Unicode.org":
\url{https://unicode.org/emoji/charts/full-emoji-list.html}.
The data are codes and descriptions of Emojis.
}
