% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{entity}
\alias{entity}
\title{Extract methods}
\usage{
entity(x, entity, ...)
}
\arguments{
\item{x}{A tweets object of the rtweet package.}

\item{entity}{A entity to extract data from.}

\item{...}{Other possible arguments currently ignored.}
}
\value{
Some information about those entities and the tweet id it comes from.
for users mentions the ids of the mentioned users are "user_id", "user_id_str"
(not "id_str")
}
\description{
Extract entities of the tweets linked to a tweet id.
}
\details{
The position of where does this occur is not provided.
}
