% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{extend_label_df}
\alias{extend_label_df}
\title{extend_label_df}
\usage{
extend_label_df(dff, label = "other")
}
\arguments{
\item{dff}{Extended data.frame within larger json
response object}

\item{label}{New label to represent the other data.frame}
}
\value{
data_frame
}
\description{
extend_label_df
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}
}

