% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections-entries.R, R/list-statuses.R
\name{f}
\alias{f}
\alias{f}
\title{GET collections/entries¶}
\usage{
f(list_id = NULL, slug = NULL, owner_user = NULL, since_id = NULL,
  max_id = NULL, count = 5000, include_rts = TRUE, token = NULL)

f(list_id = NULL, slug = NULL, owner_user = NULL, since_id = NULL,
  max_id = NULL, count = 5000, include_rts = TRUE, token = NULL)
}
\arguments{
\item{list_id}{required The numerical id of the list.}

\item{slug}{required You can identify a list by its slug instead of its numerical id. If you decide
to do so, note that you’ll also have to specify the list owner using the
owner_id or owner_screen_name parameters.}

\item{owner_user}{optional The screen name or user ID of the user who owns the list being requested by a slug.}

\item{since_id}{optional Returns results with an ID greater than (that is, more recent than) the
specified ID. There are limits to the number of Tweets which can be accessed
through the API. If the limit of Tweets has occured since the since_id, the
since_id will be forced to the oldest ID available.}

\item{max_id}{optional Returns results with an ID less than (that is, older than) or equal to the
specified ID.}

\item{count}{optional Specifies the maximum number of results to include in the response. Specify a
count between 1 and 200. A next_cursor value will be provided in the response
if additional results are available.}

\item{include_rts}{optional When set to either true , t or 1 , the list timeline will contain native
retweets (if they exist) in addition to the standard stream of tweets. The
output format of retweeted tweets is identical to the representation you see in
home_timeline.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find
instructions on how to create tokens and setup an environment
variable in the tokens vignette (in r, send \code{?tokens} to
console).}

\item{id}{required The identifier of the Collection to return results for. "539487832448843776"}

\item{max_position}{optional Returns results with a position value less than or equal to the specified
position.}

\item{min_position}{optional Returns results with a position greater than the specified position.}

\item{count}{optional Specifies the number of results to retrieve per “page.”}
}
\value{
data

data
}
\description{
Retrieve the identified Collection, presented as a list of the
  Tweets curated within. The response structure of this method
  differs significantly from timelines you may be used to
  working with in the Twitter REST API. To navigate a
  Collection, use the position object of a response, which includes
  attributes for max_position, min_position, and was_truncated.
  was_truncated indicates whether additional Tweets exist in
  the collection outside of the range of the current request.
  To retrieve Tweets further back in time, use the value of
  min_position found in the current response as the max_position
  parameter in your next call to this endpoint.

GET lists/statuses¶
}
