% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tokens.R
\name{fetch_tokens}
\alias{fetch_tokens}
\title{fetch_tokens}
\usage{
fetch_tokens(tokens, query, sleep = FALSE)
}
\arguments{
\item{tokens}{list of oauth tokens}

\item{query}{character vector, Twitter API query of interest}

\item{sleep}{logical indicating whether to force system sleep if
rate limit is exhausted. defaults to \code{sleep = FALSE}.}
}
\value{
token with non-exhausted rate limit
}
\description{
Fetch tokens based on remaining rate limit. Use this
  function to cycle through multiple tokens until rate limit
  remaining is greater than 0.
}

