% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\docType{data}
\name{Fields}
\alias{Fields}
\alias{fields}
\alias{media_fields}
\alias{place_fields}
\alias{poll_fields}
\alias{list_fields}
\alias{tweet_fields}
\alias{user_fields}
\alias{metrics_fields}
\title{Fields}
\format{
An object of class \code{character} of length 13.

An object of class \code{character} of length 8.

An object of class \code{character} of length 5.

An object of class \code{character} of length 6.

An object of class \code{character} of length 18.

An object of class \code{character} of length 15.

An object of class \code{character} of length 4.
}
\usage{
media_fields

place_fields

poll_fields

list_fields

tweet_fields

user_fields

metrics_fields
}
\description{
Arguments of expansion that select which values are returned.
Fields are possible for:
\itemize{
\item \href{https://developer.twitter.com/en/docs/twitter-api/data-dictionary/object-model/tweet}{Tweets}
\item \href{https://developer.twitter.com/en/docs/twitter-api/data-dictionary/object-model/user}{Users}
\item \href{https://developer.twitter.com/en/docs/twitter-api/data-dictionary/object-model/media}{Media}
\item \href{https://developer.twitter.com/en/docs/twitter-api/data-dictionary/object-model/poll}{Polls}
\item \href{https://developer.twitter.com/en/docs/twitter-api/data-dictionary/object-model/place}{Places}
\item \href{https://developer.twitter.com/en/docs/twitter-api/data-dictionary/object-model/lists}{Lists}
}
}
\examples{
media_fields
place_fields
poll_fields
tweet_fields
user_fields
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/fields}
}
\seealso{
\link{Expansions}, \code{\link[=set_fields]{set_fields()}}
}
\keyword{datasets}
