% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woeids.R
\name{find_woeid}
\alias{find_woeid}
\title{find_woeid}
\usage{
find_woeid(x)
}
\arguments{
\item{x}{Character string, place}
}
\description{
Returns WOEID number for desired town or country.
}
\keyword{internal}

