% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{flatten_id}
\alias{flatten_id}
\title{function to flatten IDs vars}
\usage{
flatten_id(x)
}
\arguments{
\item{x}{Data object}
}
\value{
Data with each observation of IDs represented by a single string,
  separating each ID with a space.
}
\description{
function to flatten IDs vars
}
\keyword{internal}
