% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{flatten_var}
\alias{flatten_var}
\title{function to flatten data}
\usage{
flatten_var(x)
}
\arguments{
\item{x}{Data}
}
\value{
Flatten data in preparation to be saved as CSV.
}
\description{
function to flatten data
}
\keyword{internal}
