% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{follow_user}
\alias{follow_user}
\title{follow_user}
\usage{
follow_user(user, destroy = FALSE, notify = FALSE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{destroy}{Logical indicating whether to post (add) or
remove (delete) target tweet as favorite.}

\item{notify}{Logical indicating whether to enable notifications
for target user. Defaults to false.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Follows target twitter user.
}
\examples{
\dontrun{
follow_user("BarackObama")
}
}

