% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{followers_get}
\alias{followers_get}
\title{followers_get}
\usage{
followers_get(user, token = NULL, page = "-1", stringify = TRUE)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{token}{OAuth token (1.0 or 2.0). By default
\code{token = NULL} fetches a non-exhausted token from
an environment variable.}

\item{page}{Default \code{page = -1} specifies first page of json
results. Other pages specified via cursor values supplied by
Twitter API response object.}

\item{stringify}{logical, indicating whether to return user ids
as strings (some ids are too long to be read as numeric).
Defaults to \code{TRUE}.}
}
\value{
user ids
}
\description{
Returns follower ids of specified Twitter user.
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}
}

