% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date}
\alias{format_date}
\title{format_date}
\usage{
format_date(x, date = FALSE)
}
\arguments{
\item{x}{Date or datetime vector returned by a Twitter API.}

\item{date}{Logical, indicating whether to return object of
class date. Defaults to FALSE, thus returning datetime
object (class = POSIXct).}
}
\value{
Date/datetime vector.
}
\description{
format_date
}
\examples{
\dontrun{
p <- get_timeline("potus", n = 3500)
p <- format_date(p$created_at)
head(p)
}
}

