% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{friendship_update}
\alias{friendship_update}
\title{friendship_update}
\usage{
friendship_update(user, device = FALSE, retweets = FALSE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{device}{Logical indicating whether to enable or disable
device notifications from target user behaviors. Defaults
to false.}

\item{retweets}{Logical indicating whether to enable or disable
retweets from target user behaviors. Defaults to false.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Updates friendship notifications and retweet abilities.
}

