% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{get_auth}
\alias{get_auth}
\title{Get the current authentication mechanism}
\usage{
get_auth()
}
\description{
If no authentcation has been set up for this session, will call \code{\link[=use_auth]{use_auth()}}
to set up.
}
\keyword{internal}
