% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections-entries.R
\name{get_collections}
\alias{get_collections}
\title{GET collections/list}
\usage{
get_collections(user, status_id = NULL, n = 200, cursor = NULL,
  parse = TRUE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{status_id}{Optional, the identifer of the tweet for which to return results.}

\item{n}{Maximum number of results to return.}

\item{cursor}{Page identifier of results to retrieve.}

\item{parse}{Logical indicating whether to convert response object into
nested list. Defaults to true.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find instructions on how
to create tokens and setup an environment variable in the tokens vignette
(in r, send \code{?tokens} to console).}
}
\value{
Return object converted to nested list. If status code of response
  object is not 200, the response object is returned directly.
}
\description{
Find Collections created by a specific user or containing a specific curated Tweet.
Results are organized in a cursored collection.
}
