% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized.R
\name{get_favorites_}
\alias{get_favorites_}
\title{Get favorites (vectorized)}
\usage{
get_favorites_(user, n = 100, ...)
}
\arguments{
\item{user}{Vector of user names, user IDs, or a mixture of both.}

\item{n}{Number of favorites to return per user. Defaults to 100.
Must be of length 1 or equal to length of user.}

\item{\dots}{Other arguments passed on to \code{get_favorites}.}
}
\value{
A tbl data frame of tweets data with users data attribute.
}
\description{
Returns favorites data from one or more users.
}
