% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{get_followers.default}
\alias{get_followers.default}
\title{get_followers}
\usage{
\method{get_followers}{default}(user, n = 5000, retryonratelimit = FALSE,
  page = "-1", parse = TRUE, verbose = TRUE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{n}{Number of followers to return. For max return, enter
\code{n = "all"} or \code{n = 75000} (max per token).}

\item{retryonratelimit}{The max number of followers returned every
15 minutes is 75,000. If you'd like more than that, then you can
set this to TRUE and the function will try to implement a sleep
functional call internally and iterate through until n is achieved.
Defaults to FALSE.}

\item{page}{Default \code{page = -1} specifies first page of json
results. Other pages specified via cursor values supplied by
Twitter API response object.}

\item{parse}{Logical, indicating whether to return parsed
vector or nested list (fromJSON) object. By default,
\code{parse = TRUE} saves you the time [and frustrations]
associated with disentangling the Twitter API return objects.}

\item{verbose}{Logical indicating whether or not to print messages.
Only relevant if retryonratelimit = TRUE. Defaults to TRUE, prints
sleep times and followers gathered counts.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find instructions
on how to create tokens and setup an environment variable in the
tokens vignette (in r, send \code{?tokens} to console).}
}
\value{
list of follower ids and next page value (presumably
  this would be used in loops extracting more than 75,000
  followers using either multiple tokens or by waiting out
  rate limits)
}
\description{
Returns max followers per token
}
\examples{
\dontrun{
# get ids of users following the president of the US
pres <- get_followers(user = "potus")
pres

# get ids of users following the Environmental Protection Agency
epa <- get_followers(user = "epa")
epa
}

}
\seealso{
\url{https://dev.twitter.com/overview/documentation}

Other ids: \code{\link{get_friends.default}}
}
