% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{get_friends.default}
\alias{get_friends.default}
\title{get_friends}
\usage{
\method{get_friends}{default}(users, retryonratelimit = FALSE, page = "-1",
  parse = TRUE, verbose = TRUE, token = NULL)
}
\arguments{
\item{users}{Screen name or user ID of target user from which the user IDs
of friends (accounts followed BY target user) will be retrieved.}

\item{retryonratelimit}{If you'd like to retrieve 5,000 or fewer friends for
more than 15 target users, then set \code{retryonratelimit = TRUE} and
this function will use \code{\link{Sys.sleep}} until rate limits reset and
the desired number of friend networks is retrieved. This defaults
to FALSE. See details for more info regarding possible issues with timing
misfires.}

\item{page}{Default \code{page = -1} specifies first page of json
results. Other pages specified via cursor values supplied by
Twitter API response object.}

\item{parse}{Logical, indicating whether to return parsed
vector or nested list (fromJSON) object. By default,
\code{parse = TRUE} saves you the time [and frustrations]
associated with disentangling the Twitter API return objects.}

\item{verbose}{Logical indicating whether or not to include output messages.
Defaults to TRUE.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find instructions
on how to create tokens and setup an environment variable in the
tokens vignette (in r, send \code{?tokens} to console).}
}
\value{
A tibble data frame of follower IDs (one column named "user_id").
}
\description{
Requests information from Twitter's REST API
  regarding a user's friend network (i.e., accounts followed
  by a user). To request information on followers of accounts
}
\examples{
\dontrun{
# get ids of users followed by the president of the US
pres <- get_friends(users = "potus")
pres

# get friend networks of multiple users
epa <- get_friends(users = c("jack", "epa"))
epa
}

}
\seealso{
\url{https://dev.twitter.com/overview/documentation}

Other ids: \code{\link{get_followers.default}}
}
