% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized.R
\name{get_friends_}
\alias{get_friends_}
\title{Get friends (vectorized)}
\usage{
get_friends_(user, ...)
}
\arguments{
\item{user}{Vector of user names, user IDs, or both.}

\item{\dots}{Other arguments passed along to \code{get_friends}}
}
\value{
A tbl data frame of user and friends (user_id column)
}
\description{
Returns friends for one or more users
}
