% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_friends.R
\name{get_friends_max}
\alias{get_friends_max}
\title{get_friends_max}
\usage{
get_friends_max(user_ids, tokens)
}
\arguments{
\item{user_ids}{vector of user ids}

\item{tokens}{list of OAuth tokens (1.0 or 2.0)}
}
\value{
list of friends for each of multiple users
}
\description{
Requests information from Twitter's REST API
regarding a user's friend network (i.e., accounts followed
by a user). To request information on followers of accounts.
Maximizes return from get_friends() by returning max
number of friend networks for available (selected) tokens.
}

