% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likes.R
\name{get_likes}
\alias{get_likes}
\title{get_likes}
\usage{
get_likes(user, n = 200, max_id = NULL, parse = TRUE, token = NULL, ...)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{n}{Numeric, number of tweets to return.}

\item{max_id}{Character, status_id from which returned tweets
should be older than}

\item{parse}{Logical, indicating whether to return parsed
(data.frames) or nested list (fromJSON) object. By default,
\code{parse = TRUE} saves users from the time
[and frustrations] associated with disentangling the Twitter
API return objects.}

\item{token}{OAuth token (1.0 or 2.0). By default
\code{token = NULL} fetches a non-exhausted token from
an environment variable.}

\item{\dots}{Other arguments passed on to \code{make_url}.}
}
\value{
Liked tweets data returned as a tibble data_frame for a specified
 user.
}
\description{
Returns timeline of likes from a specified
  Twitter user.
}
\examples{
\dontrun{
# get 2000 likes from Hadley Wickham's account
hadley <- get_likes("hadleywickham", n = 2000, token = twitter_tokens)

# data frame where each observation (row) is a different tweet
hadley
}
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}
}

