% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retweeters.R
\name{get_retweeters}
\alias{get_retweeters}
\title{get_retweeters}
\usage{
get_retweeters(status, page = "-1", token = NULL, stringify = TRUE)
}
\arguments{
\item{status}{Numeric id of Twitter status (tweet) of interest.}

\item{page}{Character value specifying cursor value. By default
this is set to \code{page = "-1"}, which means page 1. Value
for the next page, if it exists, will be included in the
return object.}

\item{token}{OAuth token (1.0 or 2.0). By default
\code{token = NULL} fetches a non-exhausted token from
an environment variable.}

\item{stringify}{Logical, indicating whether to return user
ids as strings (some ids are too long to be read as numeric).
Defaults to \code{TRUE}.}
}
\value{
user ids
}
\description{
Returns ids of users who retweeted a specified
  Twitter status.
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}
}

