% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retweets.R
\name{get_retweets}
\alias{get_retweets}
\title{GET statuses/retweets/:id}
\usage{
get_retweets(status_id, n = 100, parse = TRUE, ..., token = NULL)
}
\arguments{
\item{status_id}{required The numerical ID of the desired status.}

\item{n}{optional Specifies the number of records to retrieve.
Must be less than or equal to 100.}

\item{parse}{Logical indicating whether to convert the response object into
an R list. Defaults to TRUE.}

\item{...}{Other arguments used as parameters in the query sent to
Twitter's rest API, for example, \code{trim_user = TRUE}.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find
instructions on how to create tokens and setup an environment
variable in the tokens vignette (in r, send \code{?tokens} to
console).}
}
\value{
data
}
\description{
Returns a collection of the 100 most recent retweets of the Tweet
specified by the id parameter. Twitter's API is currently limited to 100
or fewer retweeters.
}
