% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorized.R
\name{get_timeline.default}
\alias{get_timeline.default}
\title{Get timeline (vectorized)}
\usage{
\method{get_timeline}{default}(user, n = 100, ...)
}
\arguments{
\item{user}{Vector of user names, user IDs, or a mixture of both.}

\item{n}{Number of tweets to return per timeline. Defaults to 100.
Must be of length 1 or equal to length of user.}

\item{\dots}{Other arguments passed on to \code{get_timeline}.}
}
\value{
A tbl data frame of tweets data with users data attribute.
}
\description{
Returns one or more user timelines.
}
