% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{get_timeline}
\alias{get_timeline}
\alias{get_my_timeline}
\alias{get_timelines}
\title{Get one or more user timelines}
\usage{
get_timeline(
  user = NULL,
  n = 100,
  max_id = NULL,
  home = FALSE,
  parse = TRUE,
  check = TRUE,
  token = NULL,
  ...
)

get_my_timeline(
  n = 100,
  max_id = NULL,
  parse = TRUE,
  check = TRUE,
  token = NULL,
  ...
)
}
\arguments{
\item{user}{Character vector of screen names or user ids.
See \code{\link[=as_screenname]{as_screenname()}} for more details.}

\item{n}{Number of tweets to return per timeline. Defaults to 100.
Must be of length 1 or equal to length of user. This number should
not exceed 3200 as Twitter limits returns to the most recent 3,200
statuses posted or retweeted by each user.}

\item{max_id}{Character, returns results with an ID less than (that is,
older than) or equal to \code{max_id}.}

\item{home}{Logical, indicating whether to return a "user" timeline
(the default, what a user has tweeted/retweeted) or a "home" timeline
(what the user would see if they logged into twitter).}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{check}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{...}{Further arguments passed on as parameters in API query.}
}
\value{
A tbl data frame of tweets data with users data attribute.
}
\description{
\code{get_timeline()} returns the timeline of any Twitter user (i.e. what they
have tweeted). \code{get_my_timeline()} returns the home timeline for the
authenticated user (i.e. the tweets you see when you log into Twitter).
}
\examples{

\dontrun{

## get most recent 3200 tweets posted by Donald Trump's account
djt <- get_timeline("realDonaldTrump", n = 3200)

## data frame where each observation (row) is a different tweet
djt

## users data for realDonaldTrump is also retrieved
users_data(djt)

## retrieve timelines of mulitple users
tmls <- get_timeline(c("KFC", "ConanOBrien", "NateSilver538"), n = 1000)

## it's returned as one data frame
tmls

## count observations for each timeline
table(tmls$screen_name)

}

}
\seealso{
\url{https://developer.twitter.com/en/docs/tweets/timelines/api-reference/get-statuses-user_timeline}

Other tweets: 
\code{\link{get_favorites}()},
\code{\link{get_mentions}()},
\code{\link{lists_statuses}()},
\code{\link{lookup_tweets}()},
\code{\link{search_tweets}()},
\code{\link{tweets_data}()},
\code{\link{tweets_with_users}()}
}
\concept{tweets}
