% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{get_trends.default}
\alias{get_trends.default}
\title{get_trends.default}
\usage{
\method{get_trends}{default}(woeid = 1, exclude = FALSE, token = NULL,
  parse = TRUE)
}
\arguments{
\item{woeid}{Numeric, WOEID (Yahoo! Where On Earth ID) or
character string of desired town or country. To browse all
available trend places, see \code{\link{trends_available}}}

\item{exclude}{Logical, indicating whether or not to exclude
hashtags}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find instructions
on how to create tokens and setup an environment variable in the
tokens vignette (in r, send \code{?tokens} to console).}

\item{parse}{Logical, indicating whether or not to parse return
trends data.}
}
\value{
Tibble data frame of trends data.
}
\description{
Returns trend data by location
}
\seealso{
Other trends: \code{\link{find_woeid}},
  \code{\link{get_trends_closest}},
  \code{\link{get_trends}}, \code{\link{trends_available}}
}
