% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{get_tweet}
\alias{get_tweet}
\title{Get tweet information}
\usage{
get_tweet(
  id,
  expansions = NULL,
  fields = NULL,
  ...,
  token = NULL,
  parse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{id}{A tweet id.}

\item{expansions}{Set \code{NULL} to not use any expansion, set \code{NA} to get all
expansions, or provide a vector with the expansions you want (create it with
\code{\link[=set_expansions]{set_expansions()}}).}

\item{fields}{Set \code{NULL} to not use any field, get all allowed fields with \code{NA},
provide a list with the fields you want (create it with \code{\link[=set_fields]{set_fields()}}).}

\item{...}{Other parameters passed to the body of the request.}

\item{token}{This endpoint accepts a OAuth2.0 authentication (can be
created via \code{\link[=rtweet_oauth]{rtweet_oauth()}}) or a bearer token (can be created via \code{\link[=rtweet_app]{rtweet_app()}}).}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}
}
\description{
Look up tweets up to 100 at the same time.
}
\examples{
if (auth_has_default()){
 get_tweet("567053242429734913", parse = FALSE)
 get_tweet(c("567053242429734913", "567053242429734913"), parse = FALSE)
 get_tweet(c("567053242429734913", "567053242429734913"),
   expansions = NULL, fields = NULL, parse = FALSE)
}
}
\references{
One tweet: \url{https://developer.twitter.com/en/docs/twitter-api/tweets/lookup/api-reference/get-tweets-id}

Multiple tweets: \url{https://developer.twitter.com/en/docs/twitter-api/tweets/lookup/api-reference/get-tweets}
}
\seealso{
\code{\link[=lookup_tweets]{lookup_tweets()}}
}
