% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{gg_cols}
\alias{gg_cols}
\title{gg_cols}
\usage{
gg_cols(n)
}
\arguments{
\item{n}{Number of colors (HEX values) to return.}
}
\value{
Character vector of HEX color codes.
}
\description{
Generates colors like those generated by
 '\pkg{ggplot2}.
}
\examples{

gg_cols(8)
}

