% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{helpers}
\alias{helpers}
\alias{urls}
\alias{hashtags}
\alias{symbols}
\alias{user_mentions}
\alias{media}
\title{Extract methods}
\usage{
urls(x)

hashtags(x)

symbols(x)

user_mentions(x)

media(x)
}
\arguments{
\item{x}{A tweets object of the rtweet package.}
}
\value{
Some information about those entities and the tweet id it comes from.
}
\description{
Extract entities of the tweets linked to a tweet id.
}
\section{Functions}{
\itemize{
\item \code{urls()}: Extract urls.

\item \code{hashtags()}: Extract hashtags.

\item \code{symbols()}: Extract symbols.

\item \code{user_mentions()}: Extract user_mentions.

\item \code{media()}: Extract media.

}}
\examples{
if (auth_has_default()) {
  statuses <- c(
    "567053242429734913",
    "266031293945503744",
    "440322224407314432"
  )

  ## lookup tweets data for given statuses
  tw <- lookup_tweets(statuses)
  urls(tw)

}
}
