% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Extract the ids}
\usage{
ids(x, ...)
}
\arguments{
\item{x}{An object of the rtweet package.}

\item{...}{Other arguments currently unused.}
}
\description{
Extract the ids of the rtweet data if present.
Depending on the object type it returns either users ids,  tweet ids or rules ids.
}
\examples{
if (auth_has_default()) {
  users <- lookup_users(c("twitter", "rladiesglobal", "_R_Foundation"))
  ids(users)
  followers <- get_followers("_R_Foundation")
  head(ids(followers))
  friends <- get_friends("_R_Foundation")
  head(ids(friends))
}
}
