% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_token.R
\name{invalidate_bearer}
\alias{invalidate_bearer}
\title{Invalidate bearer token}
\usage{
invalidate_bearer(token = NULL)
}
\arguments{
\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_use]{auth_use()}} for details.}
}
\description{
Invalidate bearer token
}
\keyword{internal}
