% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_rtweet.R
\name{join_rtweet}
\alias{join_rtweet}
\title{Joining users and tweets data}
\usage{
join_rtweet(x)
}
\arguments{
\item{x}{Data frame returned by rtweet API calls}
}
\value{
Joined users and tweets dta
}
\description{
Joining users and tweets data
}
