% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{langs}
\alias{langs}
\title{Language codes recognized by Twitter data.}
\format{A tibble with five variables and 486 observations.}
\usage{
langs
}
\description{
This data comes from the Library of Congress,
\url{http://www.loc.gov/standards/iso639-2/ISO-639-2_utf-8.txt}. The data are
descriptions and codes associated with internationally recognized languages.
Variables include translations for each language represented as
bibliographic, terminologic, alpha, english, and french.
}
\examples{
langs
}
\keyword{datasets}
