% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat_lng.R
\name{lat_long}
\alias{lat_long}
\title{lat_long}
\usage{
lat_long(x, coords = c("bbox_coords", "coords_coords", "geo_coords"))
}
\arguments{
\item{x}{Data frame returned by rtweet function.}

\item{coords}{Names of variables containing latitude and longitude coordinates.
Priority is given to bounding box coordinates (each obs consists of eight
entries) followed by the supplied order of variable names. Defaults to
c("bbox_coords", "coords_coords", "geo_coords").}
}
\value{
Returns updated data object with full information lng and lat vars.
}
\description{
Computes lat and lng variables and adds to data frame.
}
