% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_pane.R
\name{launch_rtweet_pane}
\alias{launch_rtweet_pane}
\title{Launch the RStudio Connection Pane}
\usage{
launch_rtweet_pane(token = get_tokens())
}
\arguments{
\item{token}{Your rtweet token}
}
\description{
Launch the RStudio Connection Pane
}
