% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{limits_get}
\alias{limits_get}
\title{Get api limits}
\usage{
limits_get(api = NULL)
}
\description{
If no known limits for the API has been stored for this session, \code{limits_get()} will
call \code{\link[=rate_limit]{rate_limit()}} to store them.
}
\keyword{internal}
