% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tokens.R
\name{load_tokens}
\alias{load_tokens}
\title{load_tokens}
\usage{
load_tokens(pat)
}
\value{
path character vector with path to tokens
}
\description{
Load twitter tokens from previously saved tokens
   (directed internally by previously created path locator files)
}

