% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{lookup_collections}
\alias{lookup_collections}
\title{Get collections by user or status id.}
\usage{
lookup_collections(id, n = 200, parse = TRUE, token = NULL, ...)
}
\arguments{
\item{id}{required. The identifier of the Collection to return results for
e.g., "custom-539487832448843776"}

\item{n}{Specifies the maximum number of results to include in
the response. Specify count between 1 and 200.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}

\item{...}{Other arguments passed along to composed request query.}
}
\value{
Return object converted to nested list if parsed otherwise
an HTTP response object is returned.
}
\description{
Return data for specified collection (themed grouping of Twitter statuses).
Response data varies significantly compared to most other users and tweets
data objects returned in this package.
}
\examples{

\dontrun{

## lookup a specific collection
cc <- lookup_collections("custom-539487832448843776")

## inspect data
str(cc)

}

}
