% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{lookup_friendships}
\alias{lookup_friendships}
\title{lookup_friendships}
\usage{
lookup_friendships(user, parse = TRUE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{parse}{Logical indicating whether to return parsed data frame.
Defaults to true.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find instructions
on how to create tokens and setup an environment variable in the
tokens vignette (in r, send \code{?tokens} to console).}
}
\value{
Data frame converted form returned json object. If parse is
  not true, the httr response object is returned instead.
}
\description{
Look up informaiton on friendship between authenticated
  user and up to 100 users.
}
