% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{lookup_tweets}
\alias{lookup_tweets}
\alias{lookup_statuses}
\title{Get tweets data for given statuses (status IDs).}
\usage{
lookup_tweets(
  statuses,
  parse = TRUE,
  token = NULL,
  retryonratelimit = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{statuses}{User id or screen name of target user.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{retryonratelimit}{If \code{TRUE}, and a rate limit is exhausted, will wait
until it refreshes. Most twitter rate limits refresh every 15 minutes.
If \code{FALSE}, and the rate limit is exceeded, the function will terminate
early with a warning; you'll still get back all results received up to
that point.

If you expect a query to take hours or days to perform, you should not
rely soley on \code{retryonratelimit} because it does not handle other common
failure modes like temporarily losing your internet connection.}

\item{verbose}{Show progress bars and other messages indicating current
progress?}
}
\value{
A tibble of tweets data.
}
\description{
Get tweets data for given statuses (status IDs).
}
\examples{
if (auth_has_default()) {
  statuses <- c(
    "567053242429734913",
    "266031293945503744",
    "440322224407314432"
  )
 
  tweets <- lookup_tweets(statuses)
  tweets
}
}
\seealso{
\url{https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-lookup}

Other tweets: 
\code{\link{get_favorites}()},
\code{\link{get_mentions}()},
\code{\link{get_timeline}()},
\code{\link{lists_statuses}()},
\code{\link{search_tweets}()},
\code{\link{tweets_data}()},
\code{\link{tweets_with_users}()}
}
\concept{tweets}
