% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_cursor.R
\name{max_id}
\alias{max_id}
\title{Extract minimum/maximum id from a data frame of tweets}
\usage{
max_id(x)
}
\arguments{
\item{x}{Either a data frame of tweets or a character vector of status ids.}
}
\description{
Use these functions to get earlier tweets with \code{max_id}/\code{max_id()} or later
tweets (typically tweets that have occured since your last query) with
\code{since_id}/\code{since_id()}.
}
\examples{
\dontrun{
tw <- search_tweets("#rstats")

# retrieve older tweets
older <- search_tweets("#rstats", max_id = tw)
even_older <- search_tweets("#rstats", max_id = older)

# retrieve newer tweets
newer <- search_tweets("#rstats", since_id = tw)
}
}
