% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{meta_data}
\alias{meta_data}
\title{meta_data}
\usage{
meta_data(x)
}
\arguments{
\item{x}{Data frame of class rtweet_table.}
}
\value{
List of parameters used in \code{search_users}.
}
\description{
Returns meta data search information.
}
\examples{
\dontrun{
# search for 100 tweets containing the letter r
r <- search_tweets("r")

# print tweets data (only first 10 rows are shown)
r

# extract users data
users_data(r)

# extract meta datadata
meta_data(r)
}

}

