% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtplot.R
\name{mutate.coords}
\alias{mutate.coords}
\title{mutate coords}
\usage{
mutate.coords(x)
}
\arguments{
\item{x}{Tweets data frame with at least some geo data.}
}
\value{
Data frame with long and lat columns
}
\description{
mutate coords
}

