% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R, R/rtplot.R
\name{mutate_coords}
\alias{mutate_coords}
\title{mutate_coords}
\usage{
mutate_coords(x)

mutate_coords(x)
}
\arguments{
\item{x}{Tweets data frame with some kind of geo data.}

\item{data}{Data frame generated via rtweet function.}

\item{\dots}{Args passed to points.}
}
\value{
Data frame with long and lat columns
}
\description{
Initializes rt plotting sequence

Mutates coords into numeric lat and long variables and adds
them to data frame.
}

