% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{mute_user}
\alias{mute_user}
\title{mute_user}
\usage{
mute_user(user, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Mute, or hide all content coming from, current twitter friend.
  Wrapper function for mute version of follow_user.
}

