% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-network.R
\name{network_data}
\alias{network_data}
\title{Network data}
\usage{
network_data(.x, .e = NULL)
}
\arguments{
\item{.x}{Data frame returned by rtweet function}

\item{.e}{Type of edge/link–i.e., "mention", "retweet", "quote", "reply"}
}
\value{
A from/to data edge data frame
}
\description{
Convert Twitter data into a network-friendly data frame
}
\details{
This function returns a data frame that can easily be converted to
  various network classes. For direct conversion to a network object, see
 \code{\link{network_graph}}.
}
\seealso{
network_graph
}
