% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-network.R
\name{network_graph}
\alias{network_graph}
\title{Network graph}
\usage{
network_graph(.x, .e = c("mention,retweet,reply,quote"))
}
\arguments{
\item{.x}{Data frame returned by rtweet function}

\item{.e}{Type of edge/link–i.e., "mention", "retweet", "quote", "reply".
This must be a character vector of length one or more. This value will be
split on punctuation and space (so you can include multiple types in the
same string spearated by a comma or space). The values "all" and
"semantic" are assumed to mean all edge types, which is equivalent to the
default value of \code{c("mention,retweet,reply,quote")}}
}
\value{
An igraph object
}
\description{
Convert Twitter data into network graph object (igraph)
}
\details{
This function requires previous installation of the igraph package.
  To return a network-friendly data frame, see \code{\link{network_data}}
}
\seealso{
network_data
}
