% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{next_cursor}
\alias{next_cursor}
\title{next_cursor}
\usage{
next_cursor(x)
}
\arguments{
\item{x}{Ids object (typically the output from
\link{\code{get_friends}}) or \link{\code{get_followers}}).}
}
\value{
Next cursor character string used in page argument
}
\description{
Returns next cursor from ids object.
}

