% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{next_id}
\alias{next_id}
\title{Get value for max_id}
\usage{
next_id(df)
}
\arguments{
\item{df}{Tweets data frame with "created_at" and "status_id" variables.}
}
\value{
Character string of max_id to be used in future function calls.
}
\description{
Get value for max_id
}

