% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oembed.R
\name{get_embed}
\alias{get_embed}
\alias{embed_tweet}
\title{Embed tweets}
\usage{
get_embed(screen_name, status_id, ...)

embed_tweet(screen_name, status_id, ...)
}
\arguments{
\item{screen_name, status_id}{Screename and status if of tweet to emebed.}

\item{...}{Any argument to pass to API call, 
see \href{https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-oembed.html}{official documentation} 
for the complete list.}
}
\description{
Embed tweets in Shiny or R markdown documents.
}
\details{
In R markdown use \code{\link{embed_tweet}} in your code chunk set \code{asis = TRUE}. 
In Shiny user \code{\link[shiny]{htmlOutput}}.
}
\examples{
\dontrun{
## search for 1000 tweets mentioning Hillary Clinton
hrc <- search_tweets(q = "hillaryclinton", n = 1000)

## sample tweet at random
rand <- hrc[sample(nrow(hrc), 1),]

get_embed(rand$screen_name, rand$status_id)
}

}
