% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_in_browser.R
\name{open_in_browser}
\alias{open_in_browser}
\title{Open link to tweet, user or list in the default browser}
\usage{
open_in_browser(tw, mode = "t")
}
\arguments{
\item{tw}{required The rtweet-object to open}

\item{mode}{optional Either 't', 'l' or 'u' for tweet, list or user-objects}
}
\description{
Open link to tweet, user or list in the default browser
}
\examples{

\dontrun{

open_in_browser(tw, 'u')

}
}
