% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.piper.R
\name{parse.piper}
\alias{parse.piper}
\title{parse.piper}
\usage{
parse.piper(rt, usr = TRUE)
}
\arguments{
\item{rt}{Nested list converted from json structure}

\item{usr}{Logical indicating whether to include user
obj (users data) as attribute. Defaults to true.}
}
\description{
Returns tweets data parsed via pipe
}

