% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.piper.R
\name{parse.piper.usr}
\alias{parse.piper.usr}
\title{parse.piper.usr}
\usage{
parse.piper.usr(rt, tw = FALSE)
}
\arguments{
\item{rt}{Nested list converted from json structure}

\item{tw}{Logical indicating whether to include status
obj (tweets data) as attribute. Defaults to false.}
}
\description{
Returns users data parsed via pipe
}

