% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.response.R
\name{parse_place}
\alias{parse_place}
\title{parse_place}
\usage{
parse_place(x)
}
\arguments{
\item{x}{json resposne object from user lookup Twitter
API call.}
}
\value{
data frame
}
\description{
parse_place
}
\details{
dplyr
}
\seealso{
See \url{https://dev.twitter.com/overview/
documentation} for more information on using
Twitter's API.
}

