% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_retweet}
\alias{parse_retweet}
\title{parse_retweet}
\usage{
parse_retweet(x)
}
\arguments{
\item{x}{json response object as in
\code{json_object$retweet_status}.}
}
\value{
data_frame
}
\description{
parse_retweet
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}
}

