% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3parser.R
\name{parse_statuses}
\alias{parse_statuses}
\title{parse statuses}
\usage{
parse_statuses(x)
}
\arguments{
\item{x}{List of returned Twitter data containing object statuses.}
}
\value{
Tibble data frame with users attribute.
}
\description{
parse statuses
}
