% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{parse_stream}
\alias{parse_stream}
\title{Converts Twitter stream data (JSON file) into parsed data frame.}
\usage{
parse_stream(path, ...)
}
\arguments{
\item{path}{Character, name of JSON file with data collected by
\code{\link[=stream_tweets]{stream_tweets()}}.}

\item{...}{Other arguments passed on to internal data_from_stream
function.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{jsonlite::stream_in()} instead.
}
\keyword{internal}
