% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{parse_stream_xl}
\alias{parse_stream_xl}
\title{parse_stream_xl}
\usage{
parse_stream_xl(x, by = 10000)
}
\arguments{
\item{x}{Path name for json file}

\item{by}{Number of Tweets to per chunk. By default this is set to
10,000 tweets.}
}
\value{
Data frame of tweets data
}
\description{
Returns tweets data frame from large json file
}
\details{
Reading and simplifying json files can be very slow. To
  make things more managable, \code{parse_stream_xl} does one chunk
  of Tweets at a time and then compiles the data into a data frame.
}

