% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{parse_trends}
\alias{parse_trends}
\title{parse_trends}
\usage{
parse_trends(x)
}
\arguments{
\item{x}{Nexted list fromJSON of trends data.}
}
\description{
Returns tibble data frame of trends data.
}
\keyword{internal}

