% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_tweets}
\alias{parse_tweets}
\title{parse_tweets}
\usage{
parse_tweets(x)
}
\arguments{
\item{x}{Nested json list object}
}
\value{
Tweets data as tbl (tibble) data table
}
\description{
Converts nested json list object to tweets
  data_frame
}

