% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_user}
\alias{parse_user}
\title{parse_user}
\usage{
parse_user(x)
}
\arguments{
\item{x}{json resposne object from user lookup Twitter API call.}
}
\value{
data frame
}
\description{
parse_user
}
\details{
dplyr
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation}
for more information on using Twitter's API.
}

