% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_users}
\alias{parse_users}
\title{parse_users}
\usage{
parse_users(x)
}
\arguments{
\item{x}{Nested json list object}
}
\value{
Users data as tbl (tibble) data table
}
\description{
Converts nested json list object to users
  data_frame
}

