% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.piper.R
\name{parser}
\alias{parser}
\title{Returns Parses tweets and users data}
\usage{
parser(rt, usr = TRUE, ...)
}
\arguments{
\item{rt}{Nested list converted from json structure}

\item{usr}{Logical indicating whether to include user
obj (users data) as attribute. Defaults to true.}

\item{\dots}{Other args}
}
\description{
Returns Parses tweets and users data
}

