% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3parser.R
\name{parser.statuses}
\alias{parser.statuses}
\title{parser statuses}
\usage{
\method{parser}{statuses}(x)
}
\arguments{
\item{x}{A data frame of class "statuses."}
}
\value{
A tibble data frame of tweets data.
}
\description{
Converts statuses Twitter object to tidy[er] data frame.
}
