% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3parser.R
\name{parser.user}
\alias{parser.user}
\title{parser user}
\usage{
\method{parser}{user}(x)
}
\arguments{
\item{x}{Twitter data user object or list containing users data frame.}
}
\value{
Tibble data frame.
}
\description{
Converts Twitter user object into tidy data frame.
}
