% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{plot.rt_df}
\alias{plot.rt_df}
\title{plot tweets}
\usage{
\method{plot}{rt_df}(x, ...)
}
\arguments{
\item{x}{tweets data frame object}

\item{\dots}{Other arguments passed to plot}
}
\description{
Plot tweets class data
}

