% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{plot.rtweet_df}
\alias{plot.rtweet_df}
\title{plot tweets}
\usage{
\method{plot}{rtweet_df}(x, ggplot = TRUE, ...)
}
\arguments{
\item{x}{tweets data frame object}

\item{ggplot}{Logical indicating whether to plot using \pkg{ggplot2}}

\item{\dots}{Other arguments passed to plot}
}
\description{
Plot tweets class data
}

