% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{plot.tweets}
\alias{plot.tweets}
\title{plot tweets}
\usage{
\method{plot}{tweets}(x, ...)
}
\arguments{
\item{x}{tweets object}

\item{\dots}{Other arguments passed to plot}
}
\description{
Plot tweets class data
}

