% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{plot.users}
\alias{plot.users}
\title{plot users}
\usage{
\method{plot}{users}(x, ...)
}
\arguments{
\item{x}{users object}

\item{\dots}{Other arguments passed to plot}
}
\description{
Plot users class data
}

