% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.piper.R
\name{plyget}
\alias{plyget}
\title{My version of map}
\usage{
plyget(x, f, ...)
}
\arguments{
\item{x}{Object from lhs}

\item{f}{Either atomic scalar (variable name or element position
(numbered) to be retrieved or function to be applied across
elements.}

\item{\dots}{Arguments passed along to user function}
}
\description{
My version of map
}

