% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-block.R
\name{post_block}
\alias{post_block}
\alias{post_unblock}
\title{Blocking or unblocking twitter users}
\usage{
post_block(user, unblock = FALSE, token = NULL)

post_unblock(user, token = NULL)
}
\arguments{
\item{user}{Character vector of screen names or user ids.
See \code{\link[=as_screenname]{as_screenname()}} for more details.}

\item{unblock}{Logical indicating whether to unblock the intended
friend}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{block}{Logical indicating whether to block the intended
friend}
}
\description{
\code{post_block(...)} blocks or unblocks a target twitter user.
\code{post_unblock(...)} is synonymous to \code{post_block(..., unblock=TRUE)}
}
\examples{
\dontrun{
post_block("BarackObama")
post_unblock("BarackObama")
post_block("BarackObama", unblock=TRUE) #<-same as the above
}
}
\references{
Block: \url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/mute-block-report-users/api-reference/post-blocks-create}
}
\seealso{
Other post: 
\code{\link{post_favorite}()},
\code{\link{post_follow}()},
\code{\link{post_friendship}()},
\code{\link{post_tweet}()}
}
\concept{post}
