% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_favorite}
\alias{post_favorite}
\alias{post_favourite}
\alias{favorite_tweet}
\title{Favorites target status id.}
\usage{
post_favorite(
  status_id,
  destroy = FALSE,
  include_entities = FALSE,
  token = NULL
)
}
\arguments{
\item{status_id}{Status id of target tweet.}

\item{destroy}{Logical indicating whether to post (add) or
remove (delete) target tweet as favorite.}

\item{include_entities}{Logical indicating whether to
include entities object in return.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}
}
\description{
Favorites target status id.
}
\examples{
\dontrun{
rt <- search_tweets("rstats")
r <- lapply(rt$user_id, post_favorite)
}
}
\seealso{
Other post: 
\code{\link{post_follow}()},
\code{\link{post_friendship}()},
\code{\link{post_tweet}()}
}
\concept{post}
