% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_follow}
\alias{follow_user}
\alias{post_follow}
\title{post_follow}
\usage{
post_follow(user, destroy = FALSE, mute = FALSE, notify = FALSE,
  retweets = TRUE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{destroy}{Logical indicating whether to post (add) or
remove (delete) target tweet as favorite.}

\item{mute}{Logical indicating whether to mute the intended
friend (you must already be following this account prior
to muting them)}

\item{notify}{Logical indicating whether to enable notifications
for target user. Defaults to false.}

\item{retweets}{Logical indicating whether to enable retweets
for target user. Defaults to true.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Follows target twitter user.
}
\examples{
\dontrun{
post_follow("BarackObama")
}
}
\seealso{
Other post: \code{\link{post_favorite}},
  \code{\link{post_friendship}}, \code{\link{post_mute}},
  \code{\link{post_tweet}},
  \code{\link{post_unfollow_user}}
}

