% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_friendship}
\alias{post_friendship}
\alias{friendship_update}
\title{Updates friendship notifications and retweet abilities.}
\usage{
post_friendship(user, device = FALSE, retweets = FALSE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{device}{Logical indicating whether to enable or disable
device notifications from target user behaviors. Defaults
to false.}

\item{retweets}{Logical indicating whether to enable or disable
retweets from target user behaviors. Defaults to false.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}
}
\description{
Updates friendship notifications and retweet abilities.
}
\seealso{
Other post: 
\code{\link{post_favorite}()},
\code{\link{post_follow}()},
\code{\link{post_tweet}()}
}
\concept{post}
