% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_list}
\alias{post_list}
\title{Manage Twitter lists}
\usage{
post_list(users, name, private = FALSE, description = NULL,
  destroy = FALSE, list_id = NULL, slug = NULL, token = NULL)
}
\arguments{
\item{users}{Character vectors of users to be added to list.}

\item{name}{Name of new list to create.}

\item{description}{Optional, description of list (single character string).}

\item{destroy}{Logical indicating whether to delete a list. Either `list_id` or
`slug` must be provided if `destroy = TRUE`.}

\item{list_id}{Optional, numeric ID of list.}

\item{slug}{Optional, list slug.}

\item{token}{OAuth token associated with user who owns [or will own] the list
of interest. Token must have write permissions!}
}
\value{
Response object from HTTP request.
}
\description{
Create, add users, and destroy Twitter lists
}
\examples{
\dontrun{

## CNN twitter accounts
users <- c("cnn", "cnnbrk", "cnni", "cnnpolitics", "cnnmoney",
  "cnnnewsroom", "cnnspecreport", "CNNNewsource",
  "CNNNSdigital", "CNNTonight")

## create CNN-accounts list with users and a description
cnn_lst <- post_list(users, "cnn-accounts", description = "Official CNN accounts")


}
}
