% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-message.R
\name{post_message}
\alias{post_message}
\title{Posts direct message from user's Twitter account}
\usage{
post_message(text, user, media = NULL, token = NULL)
}
\arguments{
\item{text}{Character, text of message.}

\item{user}{Screen name or user ID of message target.}

\item{media}{File path to image or video media to be
included in tweet.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}
}
\description{
Posts direct message from user's Twitter account
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/direct-messages/sending-and-receiving/api-reference/new-event}
}
