% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_unfollow_user}
\alias{post_unfollow_user}
\alias{unfollow_user}
\title{post_unfollow}
\usage{
post_unfollow_user(user, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Remove, or unfollow, current twitter friend. Wrapper function
  for destroy version of follow_user.
}
\seealso{
Other post: \code{\link{post_favorite}},
  \code{\link{post_follow}}, \code{\link{post_friendship}},
  \code{\link{post_mute}}, \code{\link{post_tweet}}
}

