% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{prep_for_csv}
\alias{prep_for_csv}
\title{Returns flattened and ID-prepended data frame, as it's done internally via
\code{save_as_csv}}
\usage{
prep_for_csv(x)
}
\description{
Returns flattened and ID-prepended data frame, as it's done internally via
\code{save_as_csv}
}
