% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_cursor.R
\name{previous_cursor}
\alias{previous_cursor}
\title{Previous cursor}
\usage{
previous_cursor(x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Reverse pagination is no longer supported.
}
\keyword{internal}
