% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{print.gcoords}
\alias{print.gcoords}
\title{print gcoords}
\usage{
\method{print}{gcoords}(x, n = 10, digits = 2)
}
\arguments{
\item{x}{Object of class coords.}

\item{n}{Number of obs to return}

\item{digits}{Number of digits to print.}
}
\description{
print gcoords
}
