% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{print.tweets}
\alias{print.tweets}
\title{print tweets}
\usage{
\method{print}{tweets}(x, ...)
}
\arguments{
\item{x}{Object of class \code{tweets}.}

\item{\dots}{Other arguments passed to \code{print}.}
}
\description{
print tweets class object
}

