% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{print.users}
\alias{print.users}
\title{print users}
\usage{
\method{print}{users}(x, ...)
}
\arguments{
\item{x}{Object of class \code{users}.}

\item{\dots}{Other arguments passed to \code{print}.}
}
\description{
print users class object
}

