% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{read_twitter_csv}
\alias{read_twitter_csv}
\title{Read comma separated value Twitter data.}
\usage{
read_twitter_csv(file, unflatten = FALSE)
}
\arguments{
\item{file}{Name of CSV file.}

\item{unflatten}{Logical indicating whether to unflatten (separate hasthags
and mentions columns on space, converting characters to lists), defaults
to FALSE.}
}
\value{
A tbl data frame of Twitter data
}
\description{
Reads Twitter data that was previously saved as a CSV file.
}
\examples{

\dontrun{

## read in data.csv
rt <- read_twitter_csv("data.csv")

}
}
\seealso{
Other datafiles: 
\code{\link{flatten}()},
\code{\link{write_as_csv}()}
}
\concept{datafiles}
