% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_v2_responses.R
\name{retrieve_errors}
\alias{retrieve_errors}
\title{Expose errors of the response}
\usage{
retrieve_errors(expr = NULL)
}
\arguments{
\item{expr}{An expression that might cause an error.
If \code{NULL} it looks for the last error.}
}
\description{
Expose errors of the response
}
\examples{
if (FALSE){
  new_rule <- stream_add_rule(list(value = "#rstats", tag = "rstats1"))
  stream_add_rule(list(value = "#rstats", tag = "rstats2")) # ERROR
  # See the full information provided by the API:
  retrieve_errors(stream_add_rule(list(value = "#rstats", tag = "rstats2")))
  retrieve_errors()
}
}
