% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retweeted_by.R
\name{retweeted_by}
\alias{retweeted_by}
\title{Retweets}
\usage{
retweeted_by(
  id,
  n = 100,
  expansions = NULL,
  fields = NULL,
  ...,
  token = NULL,
  parse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{id}{A tweet id string.}

\item{n}{Number of results to retrieve. Must be <= 100.}

\item{expansions}{Expansions you want to use see \code{\link[=tweet_expansions]{tweet_expansions()}}.
Use NULL to get all expansions, use NA to not use any field, or a vector
with the fields you want.}

\item{fields}{Fields you want to retrieve see \link{Fields}. Use NULL to get all
allowed fields, use NA to not use any field, pass a list with the fields you want.}

\item{...}{Other arguments passed to the API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}

\item{verbose}{A logical value to provide more information about paginated queries.}
}
\description{
Looks up who have retweeted a given id.
}
\examples{
rb <- retweeted_by("567053242429734913", parse = FALSE)
}
\references{
https://developer.twitter.com/en/docs/twitter-api/tweets/retweets/api-reference/get-tweets-id-retweeted_by
}
