% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rl_df}
\alias{rl_df}
\title{rl_df}
\usage{
rl_df(r)
}
\arguments{
\item{r}{rate limit response object}
}
\value{
cleaned up data frame of rate limit info
}
\description{
Returns integer values. Used for get_friends function.
}

