% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\docType{class}
\name{rt_df-class}
\alias{rt_df-class}
\title{Class "rt_df" for Tweets data}
\description{
Tweets data frame
}
\section{Class functions}{

}
\examples{

showClass("rt_df")
methods(class = "rt_df")
}
\keyword{classes}

