% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rtaes}
\alias{rtaes}
\title{rtaes}
\usage{
rtaes(x, y, color = NULL, alpha = NULL, size = NULL, shape = NULL)
}
\arguments{
\item{x}{Variable x either longitudinal coordinates or
time}

\item{y}{Variable y either latitude coordinates or
freq}

\item{color}{Variable used to determine color}

\item{alpha}{Alpha level used to set transparency}

\item{size}{Size of plot units}

\item{shape}{Shape of plot units}
}
\description{
Sets aesthetics
}

