% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rtdata}
\alias{rtdata}
\title{rtdata}
\usage{
rtdata(data, ...)
}
\arguments{
\item{data}{Data frame generated via rtweet function.}

\item{\dots}{Args passed to rtaes.}
}
\description{
Initializes rt plotting sequence
}

