% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rtline}
\alias{rtline}
\title{rtline}
\usage{
rtline(., new = FALSE, ...)
}
\arguments{
\item{.}{Data piped from rtaes function}

\item{new}{Logical indicating whether to revert to base r plot
defaults to false.}

\item{\dots}{Args passed to base plot.}
}
\description{
Plots time series line
}

