% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rtpoint}
\alias{rtpoint}
\title{rtpoint}
\usage{
rtpoint(., noise = FALSE)
}
\arguments{
\item{.}{Data piped from rtaes function}

\item{noise}{Logical indicating whether to apply minor
jitter function to points.}
}
\description{
Plots map coordinate points
}

