% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet-package.R, R/scroll.R
\docType{package}
\name{rtweet}
\alias{rtweet}
\alias{rtweet-package}
\title{rtweet}
\usage{
rtweet(url, n, parse = TRUE, ...)
}
\arguments{
\item{url}{list output object from make_url}

\item{n}{numeric, number of desired tweets to return}

\item{parse}{logical, indicating whether to return parsed or
unparsed (json) object}

\item{\dots}{named arguments passed on to httr request}
}
\description{
rtweet provides users a range of functions
  designed to extract data from Twitter's REST and
  streaming APIs.

It has three main goals:

\itemize{
\item Compose and request calls to Twitter's API
  following (the best practices for writing an API
  package) [https://cran.r-project.org/web/packages/httr/
  vignettes/api-packages.html].
\item Provide functions that can acccept multiple
  personal access tokens.
\item Make extracting data from Twitter seem a little
  easier.
}

base function.
}

