% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\docType{class}
\name{rtweet_df-class}
\alias{rtweet_df}
\alias{rtweet_df-class}
\title{rtweet_df}
\description{
Class containing rtweet [data frame] object converted
 from \code{\link{tweets}} class.
}
\examples{

showClass("rtweet_df")
methods(class = "rtweet_df")
}
\keyword{classes}

