% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{rtweet_parser}
\alias{rtweet_parser}
\title{rtweet_parser}
\usage{
rtweet_parser(x, n = NULL)
}
\arguments{
\item{x}{List, fromJSON nested list object}

\item{n}{Numeric, number of rows to return for tweets object}
}
\value{
Parsed data frame
}
\description{
Parses tweets and users objects. Returns data frames
  for each.
}

