% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{rtweet_user}
\alias{rtweet_user}
\alias{rtweet_bot}
\alias{rtweet_app}
\title{Authentication options}
\usage{
rtweet_user(api_key = NULL, api_secret = NULL)

rtweet_bot(
  api_key = askpass::askpass("Please enter your API key"),
  api_secret = askpass::askpass("Please enter your API secret"),
  access_token = askpass::askpass("Please enter your access token"),
  access_secret = askpass::askpass("Please enter your access token")
)

rtweet_app(bearer_token = askpass::askpass("Please enter your API key"))
}
\arguments{
\item{api_key, api_secret}{Application API key and secret. These are
generally not required for \code{tweet_user()} since the defaults will use
the built-in rtweet app.}

\item{access_token, access_secret}{Access token and secret.}

\item{bearer_token}{App bearer token.}
}
\description{
There are three ways that you can authenticate with the twitter API:
\itemize{
\item \code{rtweet_user()} interactively authenticates an existing twitter user.
This form is most appropriate if you want rtweet to control your
twitter account.
\item \code{rtweet_app()} authenticates as a twitter application. An application can't
perform actions (i.e. it can't tweet) but otherwise has generally higher
rate limits (i.e. you can do more searches) generally higher. See details
at \url{https://developer.twitter.com/en/docs/basics/rate-limits.html}.
This form is most appropriate if you are collecting data.
\item \code{rtweet_bot()} authenticates as bot that takes actions on behalf of an app.
This form is most appropriate if you want to create a twitter account that
is run by a computer, rather than a human.
}

To use \code{rtweet_app()} or \code{rtweet_bot()} you will need to create your own
Twitter application following the instructions in \code{vignette("auth.Rmd")}.

See \code{\link[=auth_as]{auth_as()}} to set an auth mechanism as default for this session, and
\code{\link[=auth_save]{auth_save()}} to save an auth mechanism so you can use it across multiple
sessions.
}
\section{Security}{
All of the arguments to these functions are effectively equivalent to
passwords so should ideally never be typed into the console (where they
the will be recorded in \code{.Rhistory}) or recorded in a script (which is
easy to accidentally share). Instead, the default arguments use
\code{\link[askpass:askpass]{askpass::askpass()}} to interactively prompt you for the values.
}

