% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rules}
\alias{rules}
\title{Extract the streaming rules}
\usage{
rules(x, ...)
}
\arguments{
\item{x}{An object returned by \verb{stream_*_rule}}

\item{...}{Other arguments currently ignored.}
}
\description{
Provides the information about the rules
}
\seealso{
\code{stream_add_rule()} and \code{stream_rm_rule()}.
}
