% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save Twitter data as a comma separated value file.}
\usage{
save_as_csv(x, file_name)
}
\arguments{
\item{x}{Data frame returned by rtweet function like
\code{\link{search_tweets}} to be saved as CSV file.}

\item{file_name}{Path/file name where data file is to be saved. If a ".csv"
extension isn't used, it'll be appending onto the supplied file name.}
}
\description{
Saves tweets and users data as CSV files.
}
\details{
Flattens list columns by pasting them into a single string for each
  observations. For example, a tweet that mentions four other users, for the
  mentions_user_id variable, it will include the four user IDs separated by
  a space.

It should also be noted that the save_as_csv function prepends all Twitter
  IDs (user ID and status ID) with an "x" in order to ensure spreadsheet
  programs (cough, excel, cough) don't treat the values as numeric and,
  consequently, round the numbers, preventing future identification. For this
  reason, it's best to read in data saved this way using the read_twitter_csv
  function or taking care to remove the prepended "x"s before using the
  captured ID values.
}
