% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R
\name{save_auth}
\alias{save_auth}
\title{Save an authentication mechanism for use in a future session}
\usage{
save_auth(auth, name)
}
\arguments{
\item{auth}{One of \code{\link[=auth_app]{auth_app()}}, \code{\link[=auth_bot]{auth_bot()}}, or \code{\link[=auth_user]{auth_user()}}}

\item{name}{Cache name to use}
}
\description{
You can use \code{save_auth()} and \code{\link[=use_auth]{use_auth()}} to avoid accidentally exposing
credentials in your R scripts, by caching them in a common place.
}
\examples{
\dontrun{
# save app auth for use in other sessions
save_auth(app_app("my-secret-bearer-token"), "my-app")

# later, in a different session...
use_auth("my-app")
}
}
