% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv.R
\name{save_token2renv}
\alias{save_token2renv}
\title{Saves Twitter token as an R environment variable}
\usage{
save_token2renv(token, home_dir = "~/")
}
\arguments{
\item{token}{Token object created via \code{\link{create_token}}.}

\item{home_dir}{Path to user's home directory.}
}
\description{
Saves Twitter token as an R environment variable
}
\examples{
\dontrun{

## name of app you created (replace rtweet_token with name of your app)
appname <- "rtweet_token"

## api key (example below is not a real key)
key <- "XYznzPFOFZR2a39FwWKN1Jp41"

## api secret (example below is not a real key)
secret <- "CtkGEWmSevZqJuKl6HHrBxbCybxI1xGLqrD5ynPd9jG0SoHZbD"

## create a token, storing it as object 'twitter_token'
twitter_token <- create_token(
  app = appname,
  consumer_key = key,
  consumer_secret = secret
)

## save as R environment variable
save_token2renv(twitter_token)

}

}
