% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{scopes}
\alias{scopes}
\title{Scopes of the OAuth client}
\usage{
scopes(token = NULL)
}
\arguments{
\item{token}{An authentication mechanism. Preferably set via \code{\link[=auth_as]{auth_as()}}}
}
\value{
A character with all the possible scopes.
}
\description{
Permissions given to a client of a Twitter account.
}
\examples{
scopes(NULL)
}
