% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_archive.R
\name{search_recent}
\alias{search_recent}
\title{Search recent tweets}
\usage{
search_recent(
  query,
  n = 100,
  expansions = NULL,
  fields = NULL,
  ...,
  token = NULL,
  parse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{query}{One query for matching Tweets.}

\item{expansions}{Set \code{NULL} to not use any expansion, set \code{NA} to get all
expansions, or provide a vector with the expansions you want (create it with
\code{\link[=set_expansions]{set_expansions()}}).}

\item{fields}{Set \code{NULL} to not use any field, get all allowed fields with \code{NA},
provide a list with the fields you want (create it with \code{\link[=set_fields]{set_fields()}}).}

\item{...}{Other parameters passed to the body of the request.}

\item{token}{This endpoint accepts a OAuth2.0 authentication (can be
created via \code{\link[=rtweet_oauth]{rtweet_oauth()}}) or a bearer token (can be created via \code{\link[=rtweet_app]{rtweet_app()}}).}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}
}
\description{
Look up tweets from the last seven days that match a search query.
}
\note{
OAuth2.0 requires tweet.read and users.read permissions.
}
\examples{
# sr <- search_recent("#rtweet", parse = FALSE)
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-recent}
}
