% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_users.R
\name{search_users}
\alias{search_users}
\title{Get users data on accounts identified via search query.}
\usage{
search_users(q, n = 100, parse = TRUE, token = NULL, verbose = TRUE)
}
\arguments{
\item{q}{Query to be searched, used in filtering relevant tweets to
return from Twitter's REST API. Should be a character string not
to exceed 500 characters maximum. Spaces are assumed to function
like boolean "AND" operators. To search for tweets including one
of multiple possible terms, separate search terms with spaces and
the word "OR". For example, the search \code{query = "data science"} searches for tweets using both "data" and
"science" though the words can appear anywhere and in any order
in the tweet. However, when OR is added between search terms,
\code{query = "data OR science"}, Twitter's REST API should
return any tweet that includes either "data" or "science"
appearing in the tweets. At this time, Twitter's users/search API
does not allow complex searches or queries targeting exact
phrases as is allowed by \code{search_tweets}.}

\item{n}{Numeric, specifying the total number of desired users to
return. Defaults to 100. Maximum number of users returned from a
single search is 1,000.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}

\item{verbose}{If \code{TRUE}, will display a progress bar while downloading.}
}
\value{
Data frame of users returned by query.
}
\description{
Returns data for up to 1,000 users matched by user provided search
query.
}
\examples{

\dontrun{

## search for up to 1000 users using the keyword rstats
rstats <- search_users(q = "rstats", n = 1000)

## data frame where each observation (row) is a different user
rstats

## tweets data also retrieved. can access it via tweets_data()
tweets_data(rstats)

}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/follow-search-get-users/api-reference/get-users-search}
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}

Other users: 
\code{\link{as_screenname}()},
\code{\link{lists_subscribers}()},
\code{\link{lookup_users}()},
\code{\link{tweets_with_users}()},
\code{\link{users_data}()}
}
\concept{users}
