% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\docType{methods}
\name{show,rtweet_df-method}
\alias{show,rtweet_df-method}
\title{Method show}
\usage{
\S4method{show}{rtweet_df}(object)
}
\arguments{
\item{object}{tweets object}
}
\description{
Show method for class \code{rtweet_df}
}

