% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\docType{methods}
\name{show,tweets-method}
\alias{show,tweets-method}
\title{Method show}
\usage{
\S4method{show}{tweets}(object)
}
\arguments{
\item{object}{tweets object}
}
\description{
Show method for class \code{tweets}
}

