% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\docType{methods}
\name{show,users-method}
\alias{show,users-method}
\title{Method show}
\usage{
\S4method{show}{users}(object)
}
\arguments{
\item{object}{users object}
}
\description{
Show method for class \code{users}
}

