% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stop_words}
\alias{stop_words}
\title{return stop words by language}
\usage{
stop_words(lang = NULL, n = NULL)
}
\arguments{
\item{lang}{Language to filter on must be one of "ar", "en", "es", "fr",
"in", "ja", "pt", ru", "tr", or "und".}

\item{n}{Number of stopwords to return.}
}
\value{
Character vector of stop words. These words are all lower case and contain
  no punctuation.
}
\description{
return stop words by language
}
