% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwordslangs}
\alias{stopwordslangs}
\title{stopwordslangs data}
\format{A tibble with three variables and 2,931 observations}
\usage{
stopwordslangs
}
\description{
This data comes form a group of Twitter searches conducted on 2017-09-27.
The data are commonly observed words associated with 10 different languages,
including "ar", "en", "es", "fr", "in", "ja", "pt", ru", "tr", and "und".
Variables include term (the stop words), lang, and z (standard deviation
units of logged word occurence in stopwords sample).
}
\examples{
stop_words("en")
}
\keyword{datasets}
