% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stream_params}
\alias{stream_params}
\title{stream_params}
\usage{
stream_params(stream)
}
\arguments{
\item{stream}{character stream query}
}
\value{
param character vector
}
\description{
Returns stream param.
}
\keyword{internal}

