% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream_tweets2}
\alias{stream_tweets2}
\title{Stream with hardwired reconnection method to ensure timeout integrity.}
\usage{
stream_tweets2(..., append = FALSE)
}
\arguments{
\item{...}{Args passed to \code{\link{stream_tweets}}
function. These args would likely include q (stream query),
timeout (length of time, in seconds, to maintain the connection
to the stream API), and/or file_name (name of file to save json
data as), even though function defaults technically make all
parameters optional.}

\item{append}{Logical indicating whether to append or overwrite
file_name if the file already exists. Defaults to FALSE, meaning
this function will overwrite the pre-existing file_name (in other
words, it will delete any old file with the same name as
file_name) meaning the data will be added as new lines to file if
pre-existing.}
}
\value{
Returns data as expected using original search_tweets
  function.
}
\description{
Stream with hardwired reconnection method to ensure timeout integrity.
}
