% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{tokens_sitrep}
\alias{tokens_sitrep}
\title{Twitter Tokens sitrep}
\usage{
tokens_sitrep()
}
\value{
The path of the tokens
}
\description{
Get a situation report of your current tokens useful for diagnosing problems.
}
\details{
Searches for old tokens on the user folder, if duplicate tokens are found they are deleted.
}
