% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends_available}
\alias{trends_available}
\title{Available Twitter trends along with associated WOEID.}
\usage{
trends_available(token = NULL, parse = TRUE)
}
\arguments{
\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}
}
\value{
Data frame with WOEID column. WOEID is a Yahoo! Where On
Earth ID.
}
\description{
Available Twitter trends along with associated WOEID.
}
\examples{
\dontrun{
## Retrieve available trends
trends <- trends_available()
trends

}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/trends/locations-with-trending-topics/api-reference/get-trends-available}
}
\seealso{
Other trends: 
\code{\link{get_trends}()}
}
\concept{trends}
