% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends_closest}
\alias{trends_closest}
\title{Get closest available location to given Lat, Long}
\usage{
trends_closest(lat, long, token)
}
\arguments{
\item{lat}{Latitude in degrees.}

\item{long}{Longitude in degrees.}

\item{token}{OAuth token.}
}
\description{
Uses the trends/closest API endpoint to get the
 the closest location to a given Latitude, Longitude pair. Returns
 a data frame describing the location.
}
\examples{
\dontrun{
# New York is ~ 40.7N, 74.0W
nyc_woeid <- trends_closest(40.7, -74.0, twitter_token)
}
}
