% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{trim_ts}
\alias{trim_ts}
\title{trim_ts}
\usage{
trim_ts(data, trim = 1L)
}
\arguments{
\item{data}{Time series Twitter data output by ts_data.}

\item{trim}{Number of values to drop from each time series.}
}
\value{
Trimmed data frame
}
\description{
Trim end values of time series like data.
}
