% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_data}
\alias{ts_data}
\title{Converts tweets data into time series-like data object.}
\usage{
ts_data(data, by = "days", trim = 0L)
}
\arguments{
\item{data}{Data frame or grouped data frame.}

\item{by}{Desired interval of time expressed as numeral plus secs, mins,
hours, days, weeks, months, years. If a numeric is provided, the value
is assumed to be in seconds.}

\item{trim}{Number of observations to trim off the front and end of each
time series}
}
\value{
Data frame with time, n, and grouping column if applicable.
}
\description{
Converts tweets data into time series-like data object.
}
