% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_plot.default}
\alias{ts_plot.default}
\title{ts_plot.default}
\usage{
\method{ts_plot}{default}(data, by = "days", group = NULL, ...)
}
\arguments{
\item{data}{Data frame or grouped data frame.}

\item{by}{Desired interval of time expressed as numeral plus secs,
mins, hours, days, weeks, months, years. If a numeric is
provided, the value is assumed to be in seconds.}

\item{group}{Name of grouping variable to construct multiple time
series, which still returns a data frame but it includes the
group variable as a named column.}

\item{\dots}{Passed along to trim_time. Most likely used to specify
timezone.}
}
\value{
Data frame with time, n, and grouping column if applicable.
}
\description{
Returns time series-like data frame.
}
