% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_plot}
\alias{ts_plot}
\title{ts plot}
\usage{
ts_plot(data, by = "days", ...)
}
\arguments{
\item{data}{Data frame or grouped data frame.}

\item{by}{Desired interval of time expressed as numeral plus secs,
mins, hours, days, weeks, months, years. If a numeric is
provided, the value is assumed to be in seconds.}

\item{...}{For info on all possible arguments see
\code{\link{ts_plot.default}}.}
}
\value{
Data frame with time, n, and grouping column if applicable.
}
\description{
Returns time series-like data frame.
}
