% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tsplot}
\alias{tsplot}
\title{tsplot}
\usage{
tsplot(x, by = "days", ...)
}
\arguments{
\item{x}{Tweets data frame}

\item{by}{Unit of time, e.g., \code{secs, days, weeks,
months, years}}

\item{\dots}{Other arguments passed to plot function.}
}
\description{
tsplot
}
\examples{
\dontrun{
x <- stream_tweets()
tsplot(x)
}
}

