% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtplot.R
\name{tweet.world}
\alias{tweet.world}
\title{tweet.world}
\usage{
tweet.world(x, color = NULL, bg = NULL, pch = 21, cex = 0.25,
  xlim = c(-170, 180), ylim = c(-65, 75), blank.plot = FALSE)
}
\arguments{
\item{x}{Data frame of tweets data.}

\item{color}{Color Name of variable by which dots
should be colored (grouping variable).}

\item{bg}{Background color}

\item{pch}{Point shape defaults to 21 (bubble)}

\item{cex}{Point size defaults to .25}

\item{xlim}{Min and max for x axis}

\item{ylim}{Min and max for y axis}

\item{blank.plot}{Logical inciating whether to
start from a blank plot}
}
\description{
Plots geo location tweets data on map using
latitude and longitude coordinates
}

