% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{tweet_delete}
\alias{tweet_delete}
\title{Delete tweet}
\usage{
tweet_delete(id, verbose = FALSE, token = NULL)
}
\arguments{
\item{id}{At least a tweet id.}

\item{verbose}{A logical value to provide more information about the
paginated queries (if any) and to store the data of each page.}

\item{token}{This endpoint accepts a OAuth2.0 authentication (can be
created via \code{\link[=rtweet_oauth2]{rtweet_oauth2()}}) or a bearer token (can be created via \code{\link[=rtweet_app]{rtweet_app()}}).}
}
\description{
Will delete a tweet
}
\examples{
if (FALSE) {
  # It requires Oauth authentication
  tp <- tweet_post("Running examples of #rtweet")
  td <- tweet_delete(tp$id)
}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/tweets/manage-tweets/api-reference/delete-tweets-id}
}
\seealso{
\code{\link[=tweet_post]{tweet_post()}}, \code{\link[=tweet_search_recent]{tweet_search_recent()}}, \code{\link[=user_timeline]{user_timeline()}}
}
