% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{tweet_post}
\alias{tweet_post}
\title{Post a tweet}
\usage{
tweet_post(text, ..., token = NULL)
}
\arguments{
\item{text}{Text of the tweet.}

\item{...}{Other accepted arguments.}

\item{token}{This endpoint accepts a OAuth2.0 authentication (can be
created via \code{\link[=rtweet_oauth2]{rtweet_oauth2()}}) or a bearer token (can be created via \code{\link[=rtweet_app]{rtweet_app()}}).}
}
\description{
This function uses the API v2 to post tweets.
}
\examples{
if (FALSE) {
 # It reqruires the Oauth2.0 Authentication
  tp_id <- tweet_post("Posting from #rtweet with the basic plan")
  tweet_post()
}
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/tweets/manage-tweets/api-reference/post-tweets}
}
