% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_threading.R
\name{tweet_threading}
\alias{tweet_threading}
\title{Collect statuses contained in a thread}
\usage{
tweet_threading(tw)
}
\arguments{
\item{tw}{\code{\link[rtweet]{lookup_statuses}} output containing
at least the last status in the thread}
}
\value{
\code{\link[rtweet]{lookup_statuses}} tibble
}
\description{
Return all statuses that are part of a thread
}
\details{
This function traverses up the thread until it gets to the root status
}
\examples{
tw <- lookup_statuses('1084143184664510470')
tw_thread <- tw\%>\%tweet_threading()
tw_thread
}
\seealso{
\code{\link[rtweet]{lookup_statuses}}
}
