% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_pane.R
\name{tweet_widget}
\alias{tweet_widget}
\title{Send a Tweet through the Widget}
\usage{
tweet_widget(token = get_token())
}
\arguments{
\item{token}{your rtweet token}
}
\description{
Send a Tweet through the Widget
}
