% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{tweets_data}
\alias{data_tweet}
\alias{data_tweets}
\alias{tweet_data}
\alias{tweets_data}
\title{tweets_data}
\usage{
tweets_data(users)
}
\arguments{
\item{users}{Data frame of Twitter users generated via
\code{lookup_users} or \code{search_users}.}
}
\value{
Tibble data frame of most recent tweets (if available)
  of accounts found in users data object.
}
\description{
Returns tweets data table (tibble) from users data
  object.
}
\examples{
\dontrun{
# search for 100 tweets containing the letter r
r <- search_tweets("r")

# print tweets data (only first 10 rows are shown)
r

# extract users data
users_data(r)
}

}

