% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.tweets.R
\name{tweets_df}
\alias{tweets_df}
\title{tweets_df}
\usage{
tweets_df(dat)
}
\arguments{
\item{dat}{Tweets object or nested list. Usually this is the
return object produced by \code{\link{from_js}} and
\code{\link{search_tweets}} or \code{\link{stream_tweets}}.}
}
\description{
Converts tweets object (nested list converted from
  json object) into a [tibble] data frame.
}

