% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_robust.R
\name{tweets_df_}
\alias{tweets_df_}
\title{tweets_df}
\usage{
tweets_df_(dat)
}
\arguments{
\item{dat}{Nested list parsed from Twitter's returned json object.}
}
\value{
tbl
}
\description{
tweets data
}
\details{
Returns tweets data frame.
}
