% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets_and_users.R
\name{tweets_with_users}
\alias{tweets_with_users}
\alias{users_with_tweets}
\title{Parsing data into tweets/users data tibbles}
\usage{
tweets_with_users(x)

users_with_tweets(x)
}
\arguments{
\item{x}{Unparsed data returned by rtweet API request.}
}
\value{
A tweets/users tibble (data frame) with users/tweets tibble attribute.
}
\description{
For internal use only
}
\seealso{
Other parsing: 
\code{\link{do_call_rbind}()}

Other users: 
\code{\link{as_screenname}()},
\code{\link{lists_subscribers}()},
\code{\link{lookup_users}()},
\code{\link{search_users}()},
\code{\link{users_data}()}
}
\concept{parsing}
\concept{users}
\keyword{internal}
