% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWIT.R
\name{TWIT}
\alias{TWIT}
\title{TWIT}
\usage{
TWIT(get = TRUE, url, ..., timeout = NULL, filename = NULL,
  catch_error = FALSE)
}
\arguments{
\item{get}{Locigical with the default, \code{get = TRUE},
indicating whether the provided url should be passed along via
a GET or POST request.}

\item{url}{Character vector designed to operate like
\code{parse_url()} and \code{build_url()} functions in the
httr package. The easiest way to do this is to work through
the call-specific functions as they are designed to simplify
the process. However, if one were interested in reverse-
engingeering such a thing, I would recommend checking out
\code{\link{makeurl}}.}

\item{Further}{named parameters, such as \code{token},
\code{timeout}, etc, passed on to \code{modify_url} in
httr. Unnamed parameters will be combined with
{\link{config}}.}
}
\value{
Response (json) object
}
\description{
TWIT
}
\note{
Occasionally Twitter does recommend using POST requests
  for data retrieval calls. This is usually the case when requests
  can involve long strings (containing up to 100 user_ids). For
  the most part, or at least for any function-specific requests
  (e.g., \code{get_friends}, take reflect these changes.
}

