% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{unfollow_user}
\alias{unfollow_user}
\title{unfollow_user}
\usage{
unfollow_user(user, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Remove, or unfollow, current twitter friend. Wrapper function
  for destroy version of follow_user.
}

