% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{upload_media_to_twitter}
\alias{upload_media_to_twitter}
\title{Uploads media using chunked media endpoint}
\usage{
upload_media_to_twitter(media, token)
}
\arguments{
\item{media}{Path to media file (image or movie) to upload.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Uploads media using chunked media endpoint
}
