% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uq_rtweet}
\alias{uq_rtweet}
\title{filter unique data using presumably status ID or user ID}
\usage{
uq_rtweet(x)
}
\arguments{
\item{x}{Data frame returned by rtweet functions}
}
\value{
Data frame with unique observations and users or tweets attribute data
  if already contained it
}
\description{
Removes duplicates based on ID (status ID or user ID) rather than whether each
has unique values
}
