% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{use_auth}
\alias{use_auth}
\title{Set default authentication}
\usage{
use_auth(auth = NULL)
}
\arguments{
\item{auth}{One of the following options:
\itemize{
\item \code{NULL}, the default, will look for cached authentication from a previous
session. If none are found, will automatically set up \code{\link[=auth_user]{auth_user()}};
if one is found, will use that; if multiple are found will error.
\item A string giving the name of a cached auth file.
\item An object created by \code{\link[=auth_app]{auth_app()}}, \code{\link[=auth_bot]{auth_bot()}}, or \code{\link[=auth_user]{auth_user()}}.
}}
}
\value{
Invisibly returns the previous auth mechanism.
}
\description{
\code{use_auth()} sets up the default authentication mechanism used by all
rtweet API calls. See \code{\link[=auth_user]{auth_user()}} for to learn more about the three
available authentication options.
}
\examples{
\dontrun{
# Use app auth for the remainder of this session:
use_auth(auth_app("my-secret-bearer-token"))

# Switch back to the default
use_auth()

# Load auth saved by save_auth()
use_auth("my-body")
}
}
