% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_bookmarks.R
\name{user_bookmarks}
\alias{user_bookmarks}
\title{Retrieve user bookmarks}
\usage{
user_bookmarks(
  id,
  n = 100,
  ...,
  expansions = NA,
  fields = NA,
  parse = TRUE,
  token = NULL
)
}
\arguments{
\item{id}{Twitter user id: character string identifying your account.}

\item{n}{Number of tweets to retrieve.}

\item{...}{Other arguments currently ignored.}

\item{expansions}{Expansions you want to use see \code{\link[=tweet_expansions]{tweet_expansions()}}.
Use NULL to get all expansions, use NA to not use any field, or a vector
with the fields you want.}

\item{fields}{Fields you want to retrieve see \link{Fields}. Use NULL to get all
allowed fields, use NA to not use any field, pass a list with the fields you want.}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\value{
The tweets.
}
\description{
Collects the 800 most recent bookmarked tweets of a user.
}
\note{
This endpoint requires a OAuth2.0 authentication, with tweet.read, users.read and bookmark.read permissions.
}
\examples{
if (FALSE) {
   # Requires token_oa2
   ub <- user_bookmarks("123456789", parse = FALSE, n = Inf, token = token_oa2)
}
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/tweets/bookmarks/api-reference/get-users-id-bookmarks}
}
\seealso{
\code{\link[=rtweet_oauth2]{rtweet_oauth2()}}
}
