% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.users.R
\name{user_df}
\alias{user_df}
\title{user_df}
\usage{
user_df(dat)
}
\arguments{
\item{dat}{User object or nested list. Usually this is the
return object produced by \code{\link{from_js}} and
\code{\link{lookup_users}}..}
}
\description{
Converts user object (nested list converted from
  json object) into a [tibble] data frame.
}

