% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{user_lookup}
\alias{user_lookup}
\title{user_lookup}
\usage{
user_lookup(users, token = NULL)
}
\arguments{
\item{users}{Screen name or user id of target users.}

\item{token}{OAuth token (1.0 or 2.0). By default
\code{token = NULL} fetches a non-exhausted token from
an environment variable.}
}
\value{
json response object
}
\description{
user_lookup
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}
}

