% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users-suggestions.R
\name{user_suggestions}
\alias{user_suggestions}
\title{GET users/suggestions/:slug}
\usage{
user_suggestions(slug, lang = NULL, token = NULL)
}
\arguments{
\item{slug}{required The short name of list or a category}

\item{lang}{optional Restricts the suggested categories to the requested language. The language must
be specified by the appropriate two letter ISO 639-1 representation. Currently
supported languages are provided by the GET help /
languages API request. Unsupported
language codes will receive English (en) results. If you use lang in this
request, ensure you also include it when requesting the GET users / suggestions
/ :slug list.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find
instructions on how to create tokens and setup an environment
variable in the tokens vignette (in r, send \code{?tokens} to
console).}
}
\value{
data
}
\description{
GET users/suggestions/:slug
}
