% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_timeline.R
\name{user_timeline}
\alias{user_timeline}
\title{User timeline}
\usage{
user_timeline(
  id,
  n = 100,
  expansions = NULL,
  fields = NULL,
  ...,
  token = NULL,
  parse = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{id}{A tweet id string.}

\item{n}{Number of tweets to query.}

\item{expansions}{Set \code{NULL} to not use any expansion, set \code{NA} to get all
expansions, or provide a vector with the expansions you want (create it with
\code{\link[=set_expansions]{set_expansions()}}).}

\item{fields}{Set \code{NULL} to not use any field, get all allowed fields with \code{NA},
provide a list with the fields you want (create it with \code{\link[=set_fields]{set_fields()}}).}

\item{...}{Other arguments passed to the API.}

\item{token}{These endpoints only accept a bearer token (can be created via
\code{\link[=rtweet_app]{rtweet_app()}}). In most cases you
are better of changing the default for all calls via \code{\link[=auth_as]{auth_as()}}.}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}

\item{verbose}{A logical value to provide more information about paginated queries.}
}
\description{
Looks up the timeline of a user with up to 800 tweets in the last 7 days.
}
\examples{
if (FALSE) {
  ut <- user_timeline("1599030512919650304", verbose = TRUE)
}
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/tweets/timelines/api-reference/get-users-id-mentions}
}
