% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_robust.R
\name{users_df_}
\alias{users_df_}
\title{users_df}
\usage{
users_df_(x)
}
\arguments{
\item{x}{Nested list parsed from Twitter's returned json object.}
}
\value{
tbl
}
\description{
users data
}
\details{
Returns users data frame.
}
