% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2_tweets_search_all.R
\name{v2_tweets_search_all}
\alias{v2_tweets_search_all}
\title{Search all Tweets}
\usage{
v2_tweets_search_all(
  end_time = NULL,
  expansions = NULL,
  max_results = NULL,
  media.fields = NULL,
  next_token = NULL,
  place.fields = NULL,
  poll.fields = NULL,
  query,
  since_id = NULL,
  start_time = NULL,
  tweet.fields = NULL,
  until_id = NULL,
  user.fields = NULL
)
}
\arguments{
\item{end_time}{\link{date (ISO 8601)} YYYY-MM-DDTHH:mm:ssZ (ISO 8601/RFC 3339). Used with start_time. The newest, most recent UTC timestamp to which the Tweets will be provided. Timestamp is in second granularity and is exclusive (for example, 12:00:01 excludes the first second of the minute). If used without start_time, Tweets from 30 days before end_time will be returned by default. If not specified, end_time will default to \link{now - 30 seconds}.}

\item{expansions}{\link{TODO} Comma-separated list of expansions. Expansions enable requests to expand an ID into a full object in the includes response object. Make sure to not include a space between commas and fields.}

\item{max_results}{\link{integer} The maximum number of search results to be returned by a request. A number between 10 and the system limit (currently 500). By default, a request response will return 10 results.}

\item{media.fields}{\link{TODO} Comma-separated list of additional fields to return in the media object. The response will contain the selected fields only if a Tweet contains media attachments. Make sure to not include a space between commas and fields.}

\item{next_token}{\link{string} This parameter is used to get the next 'page' of results. The value used with the parameter is pulled directly from the response provided by the API, and should not be modified. You can learn more by visiting our page on pagination.}

\item{place.fields}{\link{TODO} Comma-separated list of additional fields to return in the place object. The response will contain the selected fields only if location data is present in any of the response objects. Make sure to not include a space between commas and fields.}

\item{poll.fields}{\link{TODO} Comma-separated list of additional fields to return in the poll object. The response will contain the selected fields only if a Tweet contains a poll. Make sure to not include a space between commas and fields.}

\item{query}{\link{string} One query for matching Tweets. Up to 1024 characters.}

\item{since_id}{\link{string}    Returns results with a Tweet ID greater than (for example, more recent than) the specified ID. The ID specified is exclusive and responses will not include it. If included with the same request as a start_time parameter, only since_id will be used.}

\item{start_time}{\link{date (ISO 8601)}    YYYY-MM-DDTHH:mm:ssZ (ISO 8601/RFC 3339). The oldest UTC timestamp from which the Tweets will be provided. Timestamp is in second granularity and is inclusive (for example, 12:00:01 includes the first second of the minute). By default, a request will return Tweets from up to 30 days ago if you do not include this parameter.}

\item{tweet.fields}{\link{TODO} Comma-separated list of additional fields to return in the Tweet object. By default, the endpoint only returns id and text. Make sure to not include a space between commas and fields.}

\item{until_id}{\link{string} Returns results with a Tweet ID less than (that is, older than) the specified ID. Used with since_id. The ID specified is exclusive and responses will not include it.}

\item{user.fields}{\link{TODO} Comma-separated list of additional fields to return in the user object. By default, the endpoint does not return any user field. To use this parameter, you must include the author_id expansion parameter in the request. Make sure to not include a space between commas and fields.}
}
\value{
\itemize{
\item id \link{Default} 	string	Unique identifier of this Tweet. This is returned as a string in order to avoid complications with languages and tools that cannot handle large integers.
\item text \link{Default} 	string	The content of the Tweet. To return this field, add tweet.fields=text in the request's query parameter.
\item created_at	\link{date (ISO 8601)}	Creation time of the Tweet. To return this field, add tweet.fields=created_at in the request's query parameter.
\item \code{TODO}
}
}
\description{
Make a call to the \href{https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all}{\code{GET /2/tweets/search/all}} Twitter API endpoint. This endpoint is only available to those users who have been approved for the Academic Research product track.
}
\details{
The full-archive search endpoint returns the complete history of public Tweets matching a search query; since the first Tweet was created March 26, 2006.

The Tweets returned by this endpoint count towards the Project-level Tweet cap.
}
\section{Supported authentication}{

\itemize{
\item OAuth 2.0 Bearer token
}
}

\section{Rate limits}{

\itemize{
\item 300 requests per 15-minute window (app auth)
\item 1 request per second (app auth)
}
}

