% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{word_n}
\alias{word_n}
\title{word_n}
\usage{
word_n(data, group = NULL, ...)
}
\arguments{
\item{data}{Data frame with character column used to create word frequency
table.}

\item{group}{Optional, name of group variable. Uses group if grouped_df is
provided to data argument.}

\item{\dots}{Args passed to word_freq. Mostly used to specify name of text
column, e.g., text = "hashtags", or split = FALSE.}
}
\value{
Data frame with columns word, n, and if group is provided, then a
  third grouping column.
}
\description{
word_n
}
