% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{wordphrases}
\alias{wordphrases}
\title{phrase.tbl}
\usage{
wordphrases(txt, phrase.length = 4L, num = FALSE, excluded = NULL,
  stop.words = NULL)
}
\arguments{
\item{txt}{Tweet text vector.}

\item{phrase.length}{Number of words per phrase}

\item{num}{Logical indicating whether to exclude all numbers
from tweets. Defaults to false.}

\item{excluded}{Character string of desired text to exclude from
parsing uses regular expression.}

\item{stop.words}{optional}
}
\value{
Frequency data frame.
}
\description{
phrase.tbl
}

