% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xanyy}
\alias{\%any\%}
\alias{xanyy}
\title{xanyy}
\usage{
x \%any\% y
}
\arguments{
\item{x}{Character, name of variable of interest.}

\item{y}{Named object on which to search for \code{x}.}
}
\description{
Returns logical value indicating whether named
  object includes var name. Functions do the following:
}
\details{
\itemize{
		\item \code{\%xy\%} returns logical for each value of x
		\item \code{\%any\%} returns TRUE if y contains any of x
		\item \code{\%all\%} returns TRUE if y contains all of x
}
}

