% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xiny}
\alias{\%xy\%}
\alias{xiny}
\title{xiny}
\usage{
x \%xy\% y
}
\arguments{
\item{x}{Character, name of variable of interest.}

\item{y}{Named object on which to search for \code{x}.}
}
\value{
Logical vector of length \code{length(x)}.
}
\description{
Returns logical value indicating whether named
  object includes var name. Functions do the following:
}
\details{
\itemize{
		\item \code{\%xy\%} returns logical for each value of x
		\item \code{\%any\%} returns TRUE if y contains any of x
		\item \code{\%all\%} returns TRUE if y contains all of x
}
}
\examples{
# mpg in mtcars
"mpg" \%xy\% mtcars

# year not in mtcars
"year" \%xy\% mtcars

# check each name
c("mpg", "year") \%xy\% mtcars

# check for any
c("mpg", "year") \%any\% mtcars

# check for all
c("mpg", "year") \%all\% mtcars

}

