% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_sitrep.R
\name{auth_sitrep}
\alias{auth_sitrep}
\title{Twitter Tokens sitrep}
\usage{
auth_sitrep()
}
\value{
Invisible the path of the tokens
}
\description{
Get a situation report of your current tokens; useful for upgrading from
rtweet 0.7.0 to 1.0.0 and diagnosing problems with tokens. Called by it side effects
}
\details{
Searches for old tokens on the user folder, if duplicate tokens are found
they are deleted. Tokens are then moved to the new location.
}
\examples{
auth_sitrep()
}
