% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_token.R
\name{bearer_token}
\alias{bearer_token}
\title{Bearer token}
\usage{
bearer_token(token = NULL)
}
\arguments{
\item{token}{Oauth token created via \code{\link{create_token}}.}
}
\value{
A bearer token
}
\description{
Converts oauth token into bearer token
}
