% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_token.R
\name{bearer_token}
\alias{bearer_token}
\title{Bearer token}
\usage{
bearer_token(token = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{bearer_token()} has been deprecated because it used a rather indirect
method of generating a bearer token. Instead, use \code{\link[=rtweet_app]{rtweet_app()}}, copying
in the bearer token directly from the
\href{https://developer.twitter.com/en/portal/projects-and-apps}{Twitter developer portal}.
See \code{vignette("auth")} for full details.
}
\keyword{internal}
