% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tweets.R
\name{clean_tweets}
\alias{clean_tweets}
\title{Clean text of tweets}
\usage{
clean_tweets(x, clean = c("users", "hashtags", "urls", "media"))
}
\arguments{
\item{x}{Tweets}

\item{clean}{Type of elements to be removed.}
}
\value{
A vector with the text without the entities selected
}
\description{
Removes from the text, users mentions, hashtags, urls and media.
Some urls or other text might remain if it is not recognized as an entity by
the API.
}
\examples{
if (auth_has_default()) {
tweets <- search_tweets("weather")
tweets

# tweets
clean_tweets(tweets)
}
}
