% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{create_token}
\alias{create_token}
\title{Create custom Twitter OAuth token}
\usage{
create_token(
  app = "mytwitterapp",
  consumer_key,
  consumer_secret,
  access_token = NULL,
  access_secret = NULL,
  cache = FALSE,
  set_renv = deprecated()
)
}
\arguments{
\item{app}{Name of user created Twitter application}

\item{consumer_key}{Application API key}

\item{consumer_secret}{Application API secret User-owned
application must have \verb{Read and write} access level and
\verb{Callback URL} of \verb{http://127.0.0.1:1410}.}

\item{access_token}{Access token as supplied by Twitter (apps.twitter.com)}

\item{access_secret}{Access secret as supplied by Twitter (apps.twitter.com)}

\item{cache}{Should the token be cached in \code{\link[=token_cache_path]{token_cache_path()}}? Set this
to \code{TRUE} if you want this token to be automatically retrieved by
\code{\link[=get_token]{get_token()}}.}

\item{set_renv}{Please use \code{cache} instead.}
}
\value{
Twitter OAuth token(s) (Token1.0).
}
\description{
This function allows you to use your own Twitter app, rather than the
one built-in into rtweet. You will need to do this if you want to create
tweets or send/receive direct messages.

See \code{vignette("auth")} for more details.
}
\seealso{
\url{https://developer.twitter.com/en/docs/basics/authentication/overview/oauth}

Other tokens: 
\code{\link{get_token}()},
\code{\link{rate_limit}()}
}
\concept{tokens}
