% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{create_token}
\alias{create_token}
\title{Create custom Twitter OAuth token}
\usage{
create_token(
  app = "mytwitterapp",
  consumer_key,
  consumer_secret,
  access_token = NULL,
  access_secret = NULL,
  set_renv = TRUE
)
}
\arguments{
\item{app}{Name of user created Twitter application}

\item{consumer_key}{Application API key}

\item{consumer_secret}{Application API secret User-owned
application must have \code{Read and write} access level and
\code{Callback URL} of \code{http://127.0.0.1:1410}.}

\item{access_token}{Access token as supplied by Twitter (apps.twitter.com)}

\item{access_secret}{Access secret as supplied by Twitter (apps.twitter.com)}

\item{set_renv}{Should the token be cached?}
}
\value{
Twitter OAuth token(s) (Token1.0).
}
\description{
This function allows you to use your own Twitter app, rather than the 
one built-in into rtweet.
}
\seealso{
\url{https://developer.twitter.com/en/docs/basics/authentication/overview/oauth}

Other tokens: 
\code{\link{get_token}()},
\code{\link{rate_limit}()}
}
\concept{tokens}
