% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{create_token}
\alias{create_token}
\title{Create custom Twitter OAuth token}
\usage{
create_token(
  app = "mytwitterapp",
  consumer_key = NULL,
  consumer_secret = NULL,
  access_token = NULL,
  access_secret = NULL,
  set_renv = TRUE
)
}
\arguments{
\item{app}{Name of user created Twitter application}

\item{consumer_key, consumer_secret}{App API key and secret.}

\item{access_token, access_secret}{Access token and secret.}

\item{set_renv}{Should the token be cached?}
}
\value{
Twitter OAuth token(s) (Token1.0).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
By default, \code{create_token()} does three things: it creates an authentication
"token", sets it as the default token for the current session, and save it
to disk to make it the default for future sessions.

These three components have now been split up into three separate pieces:
use \code{\link[=rtweet_user]{rtweet_user()}}/\code{\link[=rtweet_app]{rtweet_app()}}/\code{\link[=rtweet_bot]{rtweet_bot()}} to create the token,
\code{\link[=auth_as]{auth_as()}} to make it the default for this session, and \code{\link[=auth_save]{auth_save()}} to
use it in future sessions. See \code{vignette("auth")} for full details.
}
\seealso{
Other tokens: 
\code{\link{get_token}()},
\code{\link{rate_limit}()}
}
\concept{tokens}
\keyword{internal}
