% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{data_from_stream}
\alias{data_from_stream}
\title{data_from_stream}
\usage{
data_from_stream(x, n = 10000L, n_max = -1L)
}
\arguments{
\item{x}{Character, name of json file with data collected by
\code{\link{stream_tweets}}.}

\item{n}{Number of documents (tweets) to process per interval. Defaults to 10,000.}

\item{n_max}{Number of maximum documents (tweets) to process in total. This value
is set independent from n, but it's only really useful when it's larger than n,
like when you only want to read the first million tweets from a json file that
contains 5 million tweets. Defaults to -1L, which means all lines will be read.}
}
\description{
data_from_stream
}
