% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_messages.R
\name{direct_messages_received}
\alias{direct_messages_received}
\alias{direct_messages_sent}
\title{(DEPRECATED) Get the most recent direct messages sent to the authenticating user.}
\usage{
direct_messages_received(
  since_id = NULL,
  max_id = NULL,
  n = 200,
  parse = TRUE,
  token = NULL
)

direct_messages_sent(
  since_id = NULL,
  max_id = NULL,
  n = 200,
  parse = TRUE,
  token = NULL
)
}
\value{
Return object converted to nested list. If status code of
response object is not 200, the response object is returned
directly.
}
\description{
Retrieves up to 200 of the most recently received direct messages
by the authenticating (home) user. This function requires access
token with read, write, and direct messages access.
}
\details{
Includes detailed information about the sender and
recipient user. You can request up to 200 direct messages per
call, and only the most recent 200 direct messages will be available using
this endpoint.
}
\examples{

\dontrun{

## get my direct messages
dms <- direct_messages_received()

## inspect data structure
str(dms)

## get direct messages I've sent
sdms <- direct_messages_sent()

## inspect data structure
str(dms)

}

}
\keyword{internal}
