% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emojis}
\alias{emojis}
\title{Emojis codes and descriptions data.}
\format{A tibble with two variables and 2,623 observations.}
\usage{
emojis
}
\description{
This data comes from "Unicode.org",
\url{http://unicode.org/emoji/charts/full-emoji-list.html}. The data are
codes and descriptions of Emojis.
}
\examples{
emojis
}
\keyword{datasets}
