% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expansions.R
\name{Expansions}
\alias{Expansions}
\alias{tweet_expansions}
\alias{expansions}
\alias{user_expansions}
\title{Expansions}
\usage{
tweet_expansions(attachments = TRUE, referenced_tweets = TRUE)

user_expansions()
}
\arguments{
\item{attachments}{Add attachments values? Default yes.}

\item{referenced_tweets}{Add referenced_tweets values? Default yes.}
}
\value{
A character with the characters of valid expanions.
}
\description{
Twitter parameters to add more fields on the returned values.
Main ones:
\itemize{
\item Tweet
\itemize{
\item Referenced tweets
\item Attachments
}
\item User
}
}
\examples{
tweet_expansions()
user_expansions()
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/expansions}
}
\seealso{
\link{Fields}
}
