% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.tweets.R
\name{extend_label_df}
\alias{extend_label_df}
\title{extend_label_df}
\usage{
extend_label_df(dff, label = "other")
}
\arguments{
\item{dff}{other, extended data.frame within larger json
response object}

\item{new}{label to represent the other data.frame}
}
\value{
data_frame
}
\description{
extend_label_df
}
\details{
dplyr
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation}
for more information on using Twitter's API.
}

