% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woeids.R
\name{find_woeid}
\alias{find_woeid}
\title{find_woeid}
\usage{
find_woeid(x)
}
\arguments{
\item{x}{Character string, place}
}
\description{
Returns WOEID number for desired town or country.
}
\details{
WOEID matching funcitons are offered for convenience.
  They are not nearly exhaustive and may sometimes imprecise
  (especially when multiple locations are matched to the provided
  string). To ensure validity, I recommend looking up the desired
  WOEID using one of the many different tools easily found via
  Google search.
}
\seealso{
Other trends: \code{\link{get_trends.default}},
  \code{\link{get_trends_closest}},
  \code{\link{get_trends}}, \code{\link{trends_available}}
}
\keyword{internal}
