% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{get_collections}
\alias{get_collections}
\title{Get collections by user or status id.}
\usage{
get_collections(
  user,
  status_id = NULL,
  n = 200,
  cursor = NULL,
  parse = TRUE,
  token = NULL
)
}
\arguments{
\item{user}{Screen name or user id of target user. Requests must
provide a value for one of user or status_id.}

\item{status_id}{Optional, the identifier of the tweet for which to
return results. Requests must provide a value for one of user or
status_id.}

\item{n}{Maximum number of results to return. Defaults to 200.}

\item{cursor}{Page identifier of results to retrieve. If parse = TRUE,
the next cursor value for any given request--if available--is stored
as an attribute, accessible via \code{\link[=next_cursor]{next_cursor()}}}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\value{
Return object converted to nested list if parsed otherwise
an HTTP response object is returned.
}
\description{
Find collections (themed grouping of statuses) created by specific user
or status id. Results include user, status, and collection features.
}
\examples{
if (auth_has_default()) {
  cnnc <- get_collections("cnn")
  str(cnnc$objects$users[[1]])
}
}
