% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favorites.R
\name{get_favorites}
\alias{get_favorites}
\title{get_favorites}
\usage{
get_favorites(user, n, ...)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{n}{Specifies the number of records to retrieve. Must be
less than or equal to 200; defaults to 300. 3000 is the max
number of favorites returned per token. Due to suspended or
deleted content, this function may return fewer tweets than
the desired (n) number.}

\item{...}{For other possible args see \code{\link{get_favorites.default}}.}
}
\value{
Tweets data frame.
}
\description{
Returns the most recent Tweets liked by target user.
}
\examples{
\dontrun{
# get favorites of the president of the US
pres <- get_favorites(user = "potus")
pres

# get favorites of the Environmental Protection Agency
epa <- get_favorites(user = "epa")
epa
}
}
\seealso{
Other tweets: \code{\link{get_favorites.default}},
  \code{\link{get_timeline_}},
  \code{\link{lookup_statuses}},
  \code{\link{search_tweets}}, \code{\link{stream_tweets}},
  \code{\link{tweets_data}}
}
