% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{get_followers}
\alias{get_followers}
\title{get_followers}
\usage{
get_followers(user, token = NULL, page = "-1")
}
\arguments{
\item{user}{Screen name or user id of target user}

\item{tokens}{OAuth tokens (1.0 or 2.0)}
}
\value{
user ids and page number for additional calls
}
\description{
Returns max followers per token
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation}
  for more information on using Twitter's API.
}

