% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followers.R
\name{get_followers}
\alias{get_followers}
\title{Get user IDs for accounts following target user.}
\usage{
get_followers(
  user,
  n = 5000,
  cursor = "-1",
  retryonratelimit = FALSE,
  parse = TRUE,
  verbose = TRUE,
  token = NULL,
  page = lifecycle::deprecated()
)
}
\arguments{
\item{user}{Screen name or user ID of target user from which the
user IDs of followers will be retrieved.}

\item{n}{Number of followers to return. Use \code{Inf} to download all followers.

Results are downloaded in pages of 5000, and you can download 15 pages
(i.e. 75,000 tweets) in each 15 minute period. The easiest way to download
more than that is to set \code{retryonratelimit = TRUE}.}

\item{cursor}{Which page of results to return. The default will return
the first page; can be used for manual pagination.}

\item{retryonratelimit}{If \code{TRUE}, and a rate limit is exhausted, will wait
until it refreshes. Most twitter rate limits refresh every 15 minutes.
If \code{FALSE}, and the rate limit is exceeded, the function will terminate
early with a warning; you'll still get back all results received up to
that point.

In general, if you expect a query to take hours or days to perform,
you should not rely \code{retryonratelimit} because it does not handle other
HTTP failure modes that commonly arise (i.e. you temporarily lose your
internet connection). Instead, you'll need implementing paging yourself
using \code{max_id} or \code{cursor}.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{verbose}{Show progress bars and other messages indicating current
progress?}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}

\item{page}{\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}} Please use \code{cursor} instead.}
}
\value{
A tibble data frame of follower IDs (one column named "user_id").
}
\description{
Returns a list of user IDs for the accounts following specified
user. To return more than 75,000 user IDs in a single call (the
rate limit maximum), set "retryonratelimit" to TRUE.
}
\examples{

\dontrun{

## get 5000 ids of users following the KFC account
(kfc <- get_followers("KFC"))

## get max number [per fresh token] of POTUS follower IDs
(pres <- get_followers("potus", n = 75000))
## You might use retryonratelimit = TRUE to obtain exactly 750000

}

}
\seealso{
\url{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-followers-ids}

Other ids: 
\code{\link{get_friends}()},
\code{\link{next_cursor}()}
}
\concept{ids}
