% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retweets.R
\name{get_retweeters}
\alias{get_retweeters}
\title{Getuser IDs of users who retweeted a given status.}
\usage{
get_retweeters(status_id, n = 100, parse = TRUE, token = NULL)
}
\arguments{
\item{status_id}{required The status ID of the desired status.}

\item{n}{Specifies the number of records to retrieve.
Best if intervals of 100.}

\item{parse}{Logical indicating whether to convert the response object into
an R list. Defaults to TRUE.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find
instructions on how to create tokens and setup an environment
variable in the tokens vignette (in r, send \code{?tokens} to
console).}

\item{cursor}{semi-optional Causes the list of IDs to be broken into pages
of no more than 100 IDs at a time. The number of IDs returned is not
guaranteed to be 100 as suspended users are filtered out after
connections are queried. If no cursor is provided, a
value of -1 will be assumed, which is the first "page."}
}
\value{
data
}
\description{
Getuser IDs of users who retweeted a given status.
}
\details{
At time of writing, paginaton offers no additional data. See the
  post from Pipes here: \url{https://twittercommunity.com/t/paging-is-not-possible-with-statuses-retweeters-ids-json/71298/8}
}
\seealso{
Other retweets: \code{\link{get_retweets}}
}
