% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retweets.R
\name{get_retweeters}
\alias{get_retweeters}
\title{Get user IDs of users who retweeted a given status.}
\usage{
get_retweeters(status_id, n = 100, parse = TRUE, token = NULL)
}
\arguments{
\item{status_id}{required The status ID of the desired status.}

\item{n}{Specifies the number of records to retrieve.  Best if
intervals of 100.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}
}
\value{
data
}
\description{
Returns user IDs of users who retweeted a given status. At the
current time, this function is limited in returning a maximum of
100 users for a given status.
}
\details{
At time of writing, pagination offers no additional
data.
}
\seealso{
Other retweets: 
\code{\link{get_retweets}()}
}
\concept{retweets}
