% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{get_token}
\alias{get_token}
\alias{get_tokens}
\title{Fetch Twitter OAuth token}
\usage{
get_token()

get_tokens()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link[=auth_get]{auth_get()}} instead.
}
\seealso{
Other tokens: 
\code{\link{create_token}()},
\code{\link{rate_limit}()}
}
\concept{tokens}
