% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{get_trends}
\alias{get_trends}
\title{get_trends}
\usage{
get_trends(woeid, exclude = FALSE, token = NULL)
}
\arguments{
\item{woeid}{Character, WOEID is a Yahoo! Where On
Earth ID.}

\item{exclude}{Logical, indicating whether or not to exclude
hashtags}

\item{token}{OAuth token (1.0 or 2.0). By default
\code{token = NULL} fetches a non-exhausted token from
an environment variable tokens.}
}
\value{
Trend data for a given location.
}
\description{
Returns Twitter trends
}

