% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statuses.R
\name{get_tweet}
\alias{get_tweet}
\title{Get tweet information}
\usage{
get_tweet(
  id,
  expansions = NULL,
  fields = NULL,
  ...,
  token = NULL,
  parse = TRUE
)
}
\arguments{
\item{id}{A tweet id.}

\item{expansions}{Expansions you want to use see \code{\link[=tweet_expansions]{tweet_expansions()}}.
Use NULL to get all expansions, use NA to not use any field, or a vector
with the fields you want.}

\item{fields}{Fields you want to retrieve see \link{Fields}. Use NULL to get all
allowed fields, use NA to not use any field, pass a list with the fields you want.}

\item{...}{Other parameters passed to the body of the request.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}
}
\description{
Look up tweets up to 100 at the same time.
}
\examples{
get_tweet("567053242429734913", parse = FALSE)
get_tweet(c("567053242429734913", "567053242429734913"), parse = FALSE)
get_tweet(c("567053242429734913", "567053242429734913"),
     expansions = NULL, fields = NULL, parse = FALSE)
}
\references{
One tweet: \url{https://developer.twitter.com/en/docs/twitter-api/tweets/lookup/api-reference/get-tweets-id}

Multiple tweets: \url{https://developer.twitter.com/en/docs/twitter-api/tweets/lookup/api-reference/get-tweets}
}
\seealso{
\code{\link[=lookup_tweets]{lookup_tweets()}}
}
