% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_cols}
\alias{gg_cols}
\title{ggplot colors}
\usage{
gg_cols(n)
}
\arguments{
\item{n}{Number of colors to return. If fewer than 4,
the function randomly samples 3 better colors than
what's normally produced ny the ggplot internal
mechanism.}
}
\value{
Character vector of html colors.
}
\description{
Returns ggplot2 html colors.
}

