% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entities.R
\name{helpers}
\alias{helpers}
\alias{urls}
\alias{hashtags}
\alias{symbols}
\alias{user_mentions}
\alias{media}
\title{Extract methods}
\usage{
urls(x)

hashtags(x)

symbols(x)

user_mentions(x)

media(x)
}
\arguments{
\item{x}{A tweets object of the rtweet package.}

\item{...}{Other arguments currently unused.}
}
\description{
Extract entities of the tweets.
}
\examples{
if (auth_has_default()) {
  statuses <- c(
    "567053242429734913",
    "266031293945503744",
    "440322224407314432"
  )

  ## lookup tweets data for given statuses
  tw <- lookup_tweets(statuses)
  urls(tw)

}
}
