% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_token.R
\name{invalidate_bearer}
\alias{invalidate_bearer}
\title{Invalidate bearer token}
\usage{
invalidate_bearer(token = NULL)
}
\arguments{
\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link{get_token}()}. You should only need to 
use this argument if you are wrapping rtweet in a package. 

See \code{vignette("auth")} for more details.}
}
\description{
Invalidate bearer token
}
\keyword{internal}
