% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_token.R
\name{invalidate_bearer}
\alias{invalidate_bearer}
\title{Invalidate bearer token}
\usage{
invalidate_bearer(bearer, token)
}
\arguments{
\item{token}{A bearer token}
}
\description{
Invalidate bearer token
}
\keyword{internal}
