% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_token.R
\name{invalidate_bearer}
\alias{invalidate_bearer}
\title{Invalidate bearer token}
\usage{
invalidate_bearer(token = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Invalidation of the bearer token is no longer the responsibility of rtweet.
This is something you should instead perform in the \href{https://developer.twitter.com/en/portal/projects-and-apps}{Twitter developer portal}.
}
\keyword{internal}
