% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearer_token.R
\name{invalidate_bearer}
\alias{invalidate_bearer}
\title{Invalidate bearer token}
\usage{
invalidate_bearer(token = NULL)
}
\arguments{
\item{token}{Oauth token created via \code{\link{create_token}}.}
}
\description{
Invalidate bearer token
}
\keyword{internal}
