% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{links}
\alias{links}
\title{Create the links}
\usage{
links(x, ...)
}
\arguments{
\item{x}{An object of the rtweet package.}

\item{...}{Other arguments currently unused.}
}
\description{
create the links from the rtweet data present.
Depending on the object type it returns either users links,  tweet links or rules links.
}
