% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_members.R, R/lists_memberships.R
\name{lists_members}
\alias{lists_members}
\alias{lists_memberships}
\title{Get Twitter list members (users on a given list).}
\usage{
lists_members(
  list_id = NULL,
  slug = NULL,
  owner_user = NULL,
  n = 5000,
  cursor = "-1",
  token = NULL,
  parse = TRUE,
  ...
)

lists_memberships(
  user = NULL,
  n = 200,
  cursor = "-1",
  filter_to_owned_lists = FALSE,
  token = NULL,
  parse = TRUE,
  previous_cursor = NULL
)
}
\arguments{
\item{list_id}{required The numerical id of the list.}

\item{slug}{required You can identify a list by its slug instead of
its numerical id. If you decide to do so, note that you'll also
have to specify the list owner using the owner_id or
owner_user parameters.}

\item{owner_user}{optional The screen name or user ID of the user
who owns the list being requested by a slug.}

\item{n}{Specifies the number of results to return
per page (see cursor below). For \code{list_memberships()}, the default and
max is 200 per page. Twitter technically allows up to 1,000 per page,
but above 200 frequently results in an over capacity error.
For \code{lists_members()}, the default, and max number
of users per list, is 5,000.}

\item{cursor}{optional Breaks the results into pages. Provide a
value of -1 to begin paging. Provide values as returned in the
response body's next_cursor and previous_cursor attributes to
page back and forth in the list.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{...}{Other arguments used as parameters in query composition.}

\item{user}{Character vector of screen names or user ids.
See \code{\link[=as_screenname]{as_screenname()}} for more details.}

\item{filter_to_owned_lists}{When set to true . t or 1 , will
return just lists the authenticating user owns, and the user
represented by user_id or screen_name is a member of.}

\item{previous_cursor}{If you wish to use previous cursor instead of next,
input value here to override next cursor.}
}
\value{
Either a nested list (if parsed) or an HTTP response object.
}
\description{
Due to deleted or removed lists, the returned number of memberships
is often less than the provided n value. This is a reflection of the API and
not a unique quirk of rtweet.
}
\examples{
if (auth_has_default()) {
  # You can use a list id:
  pollsters <- lists_members("105140588")
  pollsters
  
  # But in most case it'll be easier to use a user & list name
  senators <- lists_members(owner_user = "cspan", slug = "senators")
  senators
}
if (auth_has_default()) {
  # find Twitter lists that include Nate Silver
  nate_silver <- lists_memberships("NateSilver538", n = 100)
  nate_silver
}

}
\seealso{
Other lists: 
\code{\link{lists_statuses}()},
\code{\link{lists_subscribers}()},
\code{\link{lists_subscriptions}()},
\code{\link{lists_users}()}
}
\concept{lists}
