% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists-memberships.R
\name{lists_memberships}
\alias{lists_memberships}
\title{lists/memberships}
\usage{
lists_memberships(user, n = 20, cursor = "-1",
  filter_to_owned_lists = FALSE, token = NULL, parse = TRUE)
}
\arguments{
\item{user}{The user id or screen_name of the user for whom to
return results for.}

\item{n}{The amount of results to return per page. Defaults to
20. No more than 1000 results will ever be returned in a single
page.}

\item{cursor}{optional Breaks the results into pages. Provide a
value of -1 to begin paging. Provide values as returned in the
response body's next_cursor and previous_cursor attributes to
page back and forth in the list. It is recommended to always use
cursors when the method supports them. See [node:10362] for more
information.}

\item{filter_to_owned_lists}{When set to true . t or 1 , will
return just lists the authenticating user owns, and the user
represented by user_id or screen_name is a member of.}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find
instructions on how to create tokens and setup an environment
variable in the tokens vignette (in r, send \code{?tokens} to
console).}

\item{parse}{Logical indicating whether to convert the response object into
an R list. Defaults to TRUE.}
}
\value{
Either a nested list (if parsed) or an http response object.
}
\description{
Returns the lists the specified user has been added to. If user_id
or screen_name are not provided the memberships for the
authenticating user are returned.
}
