% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_memberships.R
\name{lists_memberships}
\alias{lists_memberships}
\title{Get Twitter list memberships (lists containing a given user)}
\usage{
lists_memberships(
  user = NULL,
  n = 200,
  cursor = "-1",
  filter_to_owned_lists = FALSE,
  token = NULL,
  parse = TRUE,
  previous_cursor = NULL
)
}
\arguments{
\item{user}{Character vector of screen names or user ids.
See \code{\link[=as_screenname]{as_screenname()}} for more details.}

\item{n}{Number of tweets to return per timeline. Defaults to 100.
Must be of length 1 or equal to length of user. This number should
not exceed 3200 as Twitter limits returns to the most recent 3,200
statuses posted or retweeted by each user.}

\item{cursor}{Which page of results to return. The default will return
the first page; can be used for manual pagination.}

\item{filter_to_owned_lists}{When \code{TRUE}, will return only lists that
authenticating user owns.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{previous_cursor}{If you wish to use previous cursor instead of next,
input value here to override next cursor.}
}
\description{
Due to deleted or removed lists, the returned number of memberships
is often less than the provided n value. This is a reflection of the API and
not a unique quirk of rtweet.
}
\examples{
\dontrun{

## get up to 1000 Twitter lists that include Nate Silver
ns538 <- lists_memberships("NateSilver538", n = 1000)

## view data
ns538

}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/create-manage-lists/api-reference/get-lists-memberships}
}
