% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_statuses.R
\name{lists_statuses}
\alias{lists_statuses}
\title{Get a timeline of tweets authored by members of a specified list.}
\usage{
lists_statuses(
  list_id = NULL,
  slug = NULL,
  owner_user = NULL,
  since_id = NULL,
  max_id = NULL,
  n = 200,
  include_rts = TRUE,
  parse = TRUE,
  token = NULL
)
}
\arguments{
\item{list_id}{required The numerical id of the list.}

\item{slug}{required You can identify a list by its slug instead of
its numerical id. If you decide to do so, note that you'll also have
to specify the list owner using the owner_id or owner_screen_name
parameters.}

\item{owner_user}{optional The screen name or user ID of the user
who owns the list being requested by a slug.}

\item{since_id}{optional Returns results with an ID greater than
(that is, more recent than) the specified ID. There are limits to the
number of Tweets which can be accessed through the API. If the limit
of Tweets has occurred since the since_id, the since_id will be forced
to the oldest ID available.}

\item{max_id}{optional Returns results with an ID less than (that is,
older than) or equal to the specified ID.}

\item{n}{optional Specifies the number of results to retrieve per "page."}

\item{include_rts}{optional When set to either true, t or 1,
the list timeline will contain native retweets (if they exist) in
addition to the standard stream of tweets. The output format of
retweeted tweets is identical to the representation you see in
home_timeline.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\value{
data
}
\description{
Get a timeline of tweets authored by members of a specified list.
}
\examples{
if (default_has_auth()) {
  senator_tweets <- lists_statuses(slug = "senators", owner_user = "cspan")
  senator_tweets
} 
}
\seealso{
Other lists: 
\code{\link{lists_members}()},
\code{\link{lists_subscribers}()},
\code{\link{lists_subscriptions}()},
\code{\link{lists_users}()}

Other tweets: 
\code{\link{get_favorites}()},
\code{\link{get_mentions}()},
\code{\link{get_timeline}()},
\code{\link{lookup_tweets}()},
\code{\link{search_tweets}()},
\code{\link{tweets_data}()},
\code{\link{tweets_with_users}()}
}
\concept{lists}
\concept{tweets}
