% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_subscribers.R
\name{lists_subscribers}
\alias{lists_subscribers}
\title{Get subscribers of a specified list.}
\usage{
lists_subscribers(
  list_id = NULL,
  slug = NULL,
  owner_user = NULL,
  n = 20,
  cursor = "-1",
  parse = TRUE,
  token = NULL
)
}
\arguments{
\item{list_id}{required The numerical id of the list.}

\item{slug, owner_user}{The list name (slug) and owner.}

\item{n}{Desired number of results to return. Results are downloaded
in pages when \code{n} is large; the default value will download a single
page. Set \code{n = Inf} to download as many results as possible.

The Twitter API rate limits the number of requests you can perform
in each 15 minute period. The easiest way to download more than that is
to use \code{retryonratelimit = TRUE}.

You are not guaranteed to get exactly \code{n} results back. You will get
fewer results when tweets have been deleted or if you hit a rate limit.
You will get more results if you ask for a number of tweets that's not
a multiple of page size, e.g. if you request \code{n = 150} and the page
size is 200, you'll get 200 results back.}

\item{cursor}{Which page of results to return. The default will return
the first page; can be used for manual pagination.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\description{
Get subscribers of a specified list.
}
\examples{

\dontrun{

## get subscribers of new york times politics list
rstats <- lists_subscribers(
  slug = "new-york-times-politics",
  owner_user = "nytpolitics",
  n = 1000
)

}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/create-manage-lists/api-reference/get-lists-subscribers}
}
\seealso{
Other lists: 
\code{\link{lists_members}()},
\code{\link{lists_statuses}()},
\code{\link{lists_subscriptions}()},
\code{\link{lists_users}()}

Other users: 
\code{\link{as_screenname}()},
\code{\link{lookup_users}()},
\code{\link{search_users}()}
}
\concept{lists}
\concept{users}
