% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_subscriptions.R
\name{lists_subscriptions}
\alias{lists_subscriptions}
\title{Get list subscriptions of a given user.}
\usage{
lists_subscriptions(user, n = 20, cursor = "-1", parse = TRUE, token = NULL)
}
\arguments{
\item{user}{Either the user ID or screen name of user.}

\item{n}{Specifies the number of results to return
per page (see cursor below). The default is 20, with a maximum
of 1000.}

\item{cursor}{Causes the collection of list members
to be broken into "pages" of consistent sizes (specified by
the count parameter). If no cursor is provided, a
value of -1 will be assumed, which is the first "page."
The response from the API will include a previous_cursor
and next_cursor to allow paging back and forth. See Using
cursors to navigate collections for more information.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_use]{auth_use()}} for details.}
}
\description{
Get list subscriptions of a given user.
}
\examples{

\dontrun{

## get kearneymw subscriptions
rstats <- lists_subscriptions(
  user = "kearneymw",
  n = 1000
)

}

}
\seealso{
Other lists: 
\code{\link{lists_members}()},
\code{\link{lists_statuses}()},
\code{\link{lists_subscribers}()},
\code{\link{lists_users}()}
}
\concept{lists}
