% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{load_tokens}
\alias{load_tokens}
\title{load_tokens}
\usage{
load_tokens(pat)
}
\arguments{
\item{pat}{path character vector with path to tokens}
}
\value{
loads Twitter oauth token(s)
}
\description{
Load twitter tokens from previously saved tokens
   (directed internally by previously created path locator files)
}

