% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{lookup_collections}
\alias{lookup_collections}
\alias{get_collections}
\title{Collections API}
\usage{
lookup_collections(id, n = 200, parse = TRUE, token = NULL, ...)

get_collections(
  user = NULL,
  status_id = NULL,
  n = 200,
  cursor = NULL,
  parse = TRUE,
  token = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{get_collections()} and \code{lookup_collections()} have been deprecated
since the underlying Twitter API has been deprecated:
\url{https://developer.twitter.com/en/docs/twitter-for-websites/timelines/guides/collection}.
}
\keyword{internal}
