% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{lookup_coords}
\alias{lookup_coords}
\title{lookup_coords}
\usage{
lookup_coords(address, components = NULL, box = TRUE, ...)
}
\arguments{
\item{address}{Desired location, e.g., \code{"lawrence, KS"}}

\item{components}{Unit of analysis for address e.g.,
\code{"country:US"}. Potential components include postal_code,
country, administrative_area, locality, route.}

\item{box}{Logical indicating whether to return a bounding box of
coordinates (long1, lat1, long2, lat2) or a single point (long, lat).
Defaults to TRUE. For single point, set to false.}

\item{\dots}{Additional args passed along to params portion of
http request}
}
\value{
Numeric vector with lat and long coordinates
}
\description{
Returns lat long coordinates using google geocode api
}
\examples{
\dontrun{
lookup_coords("san francisco, CA", "country:US")
}
}
