% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{lookup_friendships}
\alias{lookup_friendships}
\title{Lookup friendship information between two specified users.}
\usage{
lookup_friendships(source, target, parse = TRUE, token = NULL)
}
\arguments{
\item{source}{Screen name or user id of source user.}

\item{target}{Screen name or user id of target user.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_use]{auth_use()}} for details.}
}
\value{
Data frame converted form returned JSON object. If parse is
not true, the HTTP response object is returned instead.
}
\description{
Gets information on friendship between two Twitter users.
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/follow-search-get-users/api-reference/get-friendships-show}
}
\seealso{
Other friends: 
\code{\link{my_friendships}()}
}
\concept{friends}
