% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{lookup_friendships}
\alias{lookup_friendships}
\title{Lookup friendship information between two specified users.}
\usage{
lookup_friendships(source, target, parse = TRUE, token = NULL)
}
\arguments{
\item{source}{Screen name or user id of source user.}

\item{target}{Screen name or user id of target user.}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}

\item{token}{Use this to override authentication for
a single API call. In many cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\description{
Gets information on friendship between two Twitter users.
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/follow-search-get-users/api-reference/get-friendships-show}
}
\seealso{
Other friends: 
\code{\link{my_friendships}()}
}
\concept{friends}
