% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{lookup_friendships}
\alias{lookup_friendships}
\title{Lookup friendship information between two specified users.}
\usage{
lookup_friendships(source, target, parse = TRUE, token = NULL)
}
\arguments{
\item{source}{Screen name or user id of source user.}

\item{target}{Screen name or user id of target user.}

\item{parse}{Logical indicating whether to return parsed data frame.
Defaults to true.}

\item{token}{Every user should have their own Oauth (Twitter API) token. By
default \code{token = NULL} this function looks for the path to a saved
Twitter token via environment variables (which is what `create_token()`
sets up by default during initial token creation). For instruction on how
to create a Twitter token see the tokens vignette, i.e.,
`vignettes("auth", "rtweet")` or see \code{?tokens}.}
}
\value{
Data frame converted form returned JSON object. If parse is
  not true, the HTTP response object is returned instead.
}
\description{
Gets information on friendship between two Twitter users.
}
\seealso{
Other friends: 
\code{\link{my_friendships}()}
}
\concept{friends}
