% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{lookup_users}
\alias{lookup_users}
\title{Get Twitter users data for given users (user IDs or screen names).}
\usage{
lookup_users(users, parse = TRUE, token = NULL)
}
\arguments{
\item{users}{User id or screen name of target user.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame. 
This protects you from the vagaries of the twitter API. Use \code{FALSE} 
to return the "raw" list produced by the JSON returned from the twitter 
API.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link{get_token}()}. You should only need to 
use this argument if you are wrapping rtweet in a package. 

See \code{vignette("auth")} for more details.}
}
\value{
A tibble of users data.
}
\description{
Returns data on up to 90,000 Twitter users. To return data on more
than 90,000 users, code must be written to iterate through user IDs
whilst avoiding rate limits, which reset every 15 minutes.
}
\examples{

\dontrun{

## select one or more twitter users to lookup
users <- c(
  "potus", "hillaryclinton", "realdonaldtrump",
  "fivethirtyeight", "cnn", "espn", "twitter"
)

## get users data
usr_df <- lookup_users(users)

## view users data
usr_df

## view tweet data for these users via tweets_data()
tweets_data(usr_df)

}

}
\seealso{
\url{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-lookup}

Other users: 
\code{\link{as_screenname}()},
\code{\link{lists_subscribers}()},
\code{\link{search_users}()},
\code{\link{tweets_with_users}()},
\code{\link{users_data}()}
}
\concept{users}
