% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.users.R
\name{lookup_users}
\alias{lookup_users}
\title{lookup_users}
\usage{
lookup_users(ids, token = NULL)
}
\arguments{
\item{ids}{User id or screen name of target user.}

\item{token}{OAuth tokens (1.0 or 2.0)}
}
\value{
response object (max is 18000 per token)
}
\description{
lookup_users
}
\seealso{
See \url{https://dev.twitter.com/overview/
  documentation} for more information on using
  Twitter's API.
}

