% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{lookup_users}
\alias{lookup_users}
\title{Get Twitter users data for given users (user IDs or screen names).}
\usage{
lookup_users(users, parse = TRUE, token = NULL)
}
\arguments{
\item{users}{User id or screen name of target user.}

\item{parse}{Logical, indicating whether or not to parse
return object into data frame(s).}

\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find
instructions on how to create tokens and setup an environment
variable in the tokens vignette (in r, send \code{?tokens} to
console).}
}
\value{
A tibble of users data.
}
\description{
Returns data on up to 90,000 Twitter users. To return data on more
than 90,000 users, code must be written to iterate through user IDs
whilst avoiding rate limits, which reset every 15 minutes.
}
\examples{

\dontrun{

## select one or more twitter users to lookup
users <- c(
  "potus", "hillaryclinton", "realdonaldtrump",
  "fivethirtyeight", "cnn", "espn", "twitter"
)

## get users data
usr_df <- lookup_users(users)

## view users data
usr_df

## view tweet data for these users via tweets_data()
tweets_data(usr_df)

}

}
\seealso{
\url{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-lookup}

Other users: \code{\link{as_screenname}},
  \code{\link{lists_subscribers}},
  \code{\link{search_users}},
  \code{\link{tweets_with_users}}, \code{\link{users_data}}
}
