% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_cursor.R
\name{max_id}
\alias{max_id}
\title{Extract minimum/maximum id from a data frame of tweets}
\usage{
max_id(df)
}
\arguments{
\item{df}{A data frame of tweets.}
}
\description{
Use these functions to get earlier tweets with \code{max_id}/\code{max_id()} or later
tweets (typically tweets that have occured since your last query) with
\code{since_id}/\code{since_id()}.
}
\examples{
\dontrun{
tw <- search_tweets("#rstats", n = 100)

# retrieve older tweets
older <- search_tweets("#rstats", n = 100, max_id = max_id(tw))
even_older <- search_tweets("#rstats", n = 100, max_id = max_id(older))

# retrieve newer tweets
newer <- search_tweets("#rstats", n = 100, since_id = since_id(tw))
}
}
