% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_cursor.R
\name{max_id}
\alias{max_id}
\title{Extract min/max id (for id based pagination)}
\usage{
max_id(x)
}
\arguments{
\item{x}{Either a data frame of tweets or a character vector of status ids.}
}
\description{
These internal helpers extract the ids passed on to the \code{max_id}
and \code{since_id} arguments to functions that use \code{\link[=TWIT_paginate_max_id]{TWIT_paginate_max_id()}}.
}
\keyword{internal}
