% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mutate_coords}
\alias{mutate_coords}
\title{mutate_coords}
\usage{
mutate_coords(data, ...)
}
\arguments{
\item{data}{Data frame generated via rtweet function.}

\item{\dots}{Args passed to points.}
}
\description{
Initializes rt plotting sequence
}

