% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.R
\name{my_friendships}
\alias{my_friendships}
\title{Lookup friendship information between users.}
\usage{
my_friendships(user, parse = TRUE, token = NULL)
}
\arguments{
\item{user}{Screen name or user id of target user.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link[=get_token]{get_token()}}. You should only need to
use this argument if you are wrapping rtweet in a package.

See \code{vignette("auth")} for more details.}
}
\value{
Data frame converted form returned JSON object. If parse is
not true, the HTTP response object is returned instead.
}
\description{
Gets information on friendship between authenticated user and up
to 100 other users.
}
\seealso{
Other friends: 
\code{\link{lookup_friendships}()}
}
\concept{friends}
