% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-network.R
\name{network_graph}
\alias{network_graph}
\title{Network graph}
\usage{
network_graph(.x, .e = NULL)
}
\arguments{
\item{.x}{Data frame returned by rtweet function}

\item{.e}{Type of edge/link–i.e., "mention", "retweet", "quote", "reply"}
}
\value{
An igraph object
}
\description{
Convert Twitter data into network graph object (igraph)
}
\details{
This function requires previous installation of the igraph package.
  To return a network-friendly data frame, see \code{\link{network_data}}
}
\seealso{
network_data
}
