% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_cursor.R
\name{next_cursor}
\alias{next_cursor}
\alias{since_id}
\title{Extract cursor (for cursor based pagination)}
\usage{
next_cursor(x)

since_id(x)
}
\description{
This internal helper extracts the cursor from the object passed to the
\code{cursor} argument of the functions that use \code{\link[=TWIT_paginate_cursor]{TWIT_paginate_cursor()}}.
}
\examples{
if (auth_has_default()) {
page1 <- get_followers("_R_Foundation")
page2 <- get_followers("_R_Foundation", cursor = page1)
}
}
\keyword{internal}
