% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_all_tweets}
\alias{parse_all_tweets}
\title{parse_all_tweets}
\usage{
parse_all_tweets(x)
}
\arguments{
\item{x}{json response object from tweet/status
Twitter API request.}
}
\value{
data_frame
}
\description{
Converts Twitter status object to neat data_frame.
  This function subsumes the parse_status function and also
  converts place and user objects.
}
\seealso{
\url{https://dev.twitter.com/overview/documentation}
}

