% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.piper.R
\name{parse_data}
\alias{parse_data}
\title{parse_data}
\usage{
parse_data(rt, tw = TRUE)
}
\arguments{
\item{rt}{Nested list converted from json structure}

\item{tw}{Logical indicating whether to include user
obj (users data) as attribute. Defaults to true. If
was generated from users-oriented function, e.g.,
lookup_users(), set to false.}
}
\value{
Data frame of tweets or users data with the other (tweets/users)
  included as an attribute. If something breaks during the parsing process,
  this function has a fail safe that will return a list (presumably with
  unbalanced lengths).
}
\description{
Returns Parses tweets and users data
}
