% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{parse_stream}
\alias{parse_stream}
\title{parse_stream}
\usage{
parse_stream(file_name)
}
\arguments{
\item{file_name}{name of file to be parsed. NOTE: if file
was created via \code{\link{stream_tweets}}, then it will
end in ".json" (see example below)}
}
\value{
Parsed tweets data with users data attribute.
}
\description{
parse_stream
}
\examples{
\dontrun{
stream_tweets(q = "", file_name = "tw", parse = FALSE)
tw <- parse_stream("tw.json")
tw
}
}

