% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{parse_stream}
\alias{parse_stream}
\title{Converts Twitter stream data (JSON file) into parsed data frame.}
\usage{
parse_stream(path, ...)
}
\arguments{
\item{path}{Character, name of JSON file with data collected by
\code{\link{stream_tweets}}.}

\item{...}{Other arguments passed on to internal data_from_stream
function.}
}
\value{
A tbl of tweets data with attribute of users data
}
\description{
Converts Twitter stream data (JSON file) into parsed data frame.
}
\examples{
\dontrun{
## run and save stream to JSON file
stream_tweets(
  "the,a,an,and", timeout = 60,
  file_name = "theaanand.json",
  parse = FALSE
)

## parse stream file into tibble data frame
rt <- parse_stream("theaanand.json")
}
}
\seealso{
Other stream tweets: 
\code{\link{stream_tweets}()}
}
\concept{stream tweets}
