% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3parser.R
\name{parser}
\alias{parser}
\title{parser method}
\usage{
parser(x, ...)
}
\arguments{
\item{x}{Data object.}

\item{\dots}{Args passed along to various methods. Basically, it's used to
specify the desired variable to extract.}
}
\value{
Parsed data object.
}
\description{
Wrangling methods for converting nested lists (supplied by Twitter)
  into data frames
}
\examples{
\dontrun{
rt <- search_tweets("lang:en", n = 500, parse = FALSE)
df <- parser(rt)
df
users_data(df)
}
}
