% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll.R
\name{parser}
\alias{parser}
\title{parser}
\usage{
parser(x, n = NULL)
}
\arguments{
\item{x}{List, fromJSON nested list object}

\item{n}{Numeric, number of desired tweets to return}
}
\description{
Parses tweets and users objects. Returns data frames
  for each.
}

