% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.piper.R
\name{parser}
\alias{parser}
\title{parser}
\usage{
parser(rt, att = TRUE)
}
\arguments{
\item{rt}{Nested list converted from json structure}

\item{att}{Logical indicating whether to include user
obj (users data) as attribute if tweets data provided to
rt argument or tweets obj (tweets data) as attribute if
users data provided to rt. Defaults to true.}
}
\description{
Returns Parses tweets and users data
}

