% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parser}
\alias{parser}
\title{parser}
\usage{
parser(x, n = NULL)
}
\arguments{
\item{x}{nested list of API data returned from fromJSON}

\item{n}{desired number to return}
}
\description{
parser
}
\keyword{internal}

