% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{plain_tweets}
\alias{plain_tweets}
\title{plain text tweets}
\usage{
plain_tweets(txt, tokenize = FALSE)
}
\arguments{
\item{txt}{Tweet text, character vector.}

\item{tokenize}{Logical indicating whether to split each tweet into
a vector of individual words. Defaults to FALSE.}
}
\value{
Stripped and cleaned up version of tweet text.
}
\description{
plain text tweets
}
