% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plain_tweets.R
\name{plain_tweets}
\alias{plain_tweets}
\title{Clean up character vector (tweets) to more of a plain text.}
\usage{
plain_tweets(x)
}
\arguments{
\item{x}{The desired character vector or data frame/list with named column/element
"text" to be cleaned and processed.}
}
\value{
Data reformatted with ascii encoding and normal ampersands and
  without URL links, line breaks, fancy spaces/tabs, fancy apostrophes,
}
\description{
Clean up character vector (tweets) to more of a plain text.
}
