% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-user.R
\name{post_friendship}
\alias{post_friendship}
\alias{friendship_update}
\title{Updates friendship notifications and retweet abilities.}
\usage{
post_friendship(user, device = FALSE, retweets = FALSE, token = NULL)
}
\arguments{
\item{user}{Character vector of screen names or user ids.
See \code{\link[=as_screenname]{as_screenname()}} for more details.}

\item{device}{Logical indicating whether to enable or disable
device notifications from target user behaviors. Defaults
to false.}

\item{retweets}{Logical indicating whether to enable or disable
retweets from target user behaviors. Defaults to false.}

\item{token}{Use this to override authentication for
a single API call. In many cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\description{
Updates friendship notifications and retweet abilities.
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/follow-search-get-users/api-reference/post-friendships-update}
}
\seealso{
Other post: 
\code{\link{post_favorite}()},
\code{\link{post_follow}()},
\code{\link{post_tweet}()}
}
\concept{post}
