% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-message.R
\name{post_message}
\alias{post_message}
\title{Posts direct message from user's Twitter account}
\usage{
post_message(text, user, media = NULL, token = NULL)
}
\arguments{
\item{text}{Character, text of message.}

\item{user}{Screen name or user ID of message target.}

\item{media}{File path to image or video media to be
included in tweet.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_use]{auth_use()}} for details.}
}
\description{
Posts direct message from user's Twitter account
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/direct-messages/sending-and-receiving/api-reference/new-event}
}
