% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-message.R
\name{post_message}
\alias{post_message}
\title{Posts direct message from user's Twitter account}
\usage{
post_message(text, user, media = NULL, token = NULL)
}
\arguments{
\item{text}{Character, text of message.}

\item{user}{Character vector of screen names or user ids.
See \code{\link[=as_screenname]{as_screenname()}} for more details.}

\item{media}{File path to image or video media to be
included in tweet.}

\item{token}{Use this to override authentication for
a single API call. In many cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}
}
\description{
Posts direct message from user's Twitter account
}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/direct-messages/sending-and-receiving/api-reference/new-event}
}
