% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_message}
\alias{post_message}
\title{Posts direct message from user's Twitter account}
\usage{
post_message(text, user, media = NULL, token = NULL)
}
\arguments{
\item{text}{Character, text of message.}

\item{user}{Screen name or user ID of message target.}

\item{media}{File path to image or video media to be
included in tweet.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Posts direct message from user's Twitter account
}
