% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.R
\name{post_tweet}
\alias{post_tweet}
\title{post_tweet}
\usage{
post_tweet(status = "my first rtweet #rstats", token = NULL)
}
\arguments{
\item{status}{Character, tweet status. Must be 140
characters or less.}

\item{token}{OAuth token. By default \code{token = NULL}
fetches a non-exhausted token from an environment
variable tokens.}
}
\description{
Posts status update to user's Twitter account
}
\examples{
\dontrun{
post_tweet("my first rtweet #rstats")
}
}

