% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{print.tweets}
\alias{print.tweets}
\title{print tweets}
\usage{
\method{print}{tweets}(x, ...)
}
\arguments{
\item{x}{tweets class object}

\item{\dots}{Other arguments passed to print}
}
\description{
print tweets class object
}

