% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rate_limit}
\alias{rate_limit}
\title{rate_limit}
\usage{
rate_limit(token, query = NULL, rest = TRUE)
}
\arguments{
\item{token}{An OAuth token (1.0 or 2.0)}

\item{query}{If null, returns entire rate limit request object as
data frame. otherwise, query returns specific values matching
the query of interest; e.g., \code{query = "lookup"} returns
remaining limit for user lookup requests;
\code{type = "followers"} returns remaining limit for
follower id requests; \code{type = "friends"} returns
remaining limit for friend id requests.}
}
\value{
response Rate limit response object or specific value of
  remaining requests
}
\description{
rate_limit
}
\seealso{
See \url{https://dev.twitter.com/overview/documentation}
  for more information on using Twitter's API.
}

