% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4.R
\name{rt_data}
\alias{rt_data}
\title{rt_data}
\usage{
rt_data(object)
}
\arguments{
\item{object}{Tweets class list object}
}
\description{
Convert tweets and users objects
  to data frames
}
\examples{
\dontrun{
tw <- search_tweets("twitter")
tw <- rt_data(tw)
tw
}
}
\keyword{classes}

