% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rtdata}
\alias{rtdata}
\title{rtdata}
\usage{
rtdata(data, by = NULL, ...)
}
\arguments{
\item{data}{Data frame generated via rtweet function.}

\item{by}{Unit of time for aggregating tweet frequency for
time series plot. Defaults to NULL. To plot as time series
by argument is required (e.g., \code{by = "days"}.}

\item{\dots}{Args passed to aes and other functions.}
}
\description{
Initializes rt plotting sequence
}

