% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet-package.R
\docType{package}
\name{rtweet-package}
\alias{rtweet}
\alias{rtweet-package}
\title{rtweet: Collecting Twitter data}
\description{
rtweet provides users a range of functions designed to extract data
from Twitter's REST and streaming APIs.
}
\details{
It has three main goals:

\itemize{
\item Formulate and send requests to Twitter's REST and stream APIs.
\item Retrieve and iterate over returned data.
\item Wrangling data into tidy structures.
}
}
\examples{
\dontrun{
## for instructions on access tokens, see the tokens vignette
vignette("auth")

## for a quick demo check the rtweet vignette
vignette("rtweet")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=rtweet}
  \item Report bugs at \url{https://github.com/mkearney/rtweet/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu}

}
