% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet.R
\docType{package}
\name{rtweet}
\alias{rtweet}
\alias{rtweet-package}
\title{rtweet}
\description{
rtweet provides users a range of functions
  designed to extract data from Twitter's REST and
  streaming APIs.

It has three main goals:

\itemize{
\item Compose and request calls to Twitter's API
  following (the best practices for writing an API
  package) [https://cran.r-project.org/web/packages/httr/
  vignettes/api-packages.html].
\item Provide functions that can acccept multiple
  personal access tokens.
\item Make extracting data from Twitter seem a little
  easier.
}
}

