% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtweet-package.R
\docType{package}
\name{rtweet}
\alias{rtweet}
\alias{rtweet-package}
\alias{rtweets}
\alias{rtwitter}
\alias{tokens}
\title{rtweet}
\description{
rtweet provides users a range of functions
  designed to extract data from Twitter's REST and
  streaming APIs.

It has three main goals:

\itemize{
\item Formulate and send requests to Twitter's REST and stream APIs.
\item Retrieve and iterate over returned data.
\item Wrangling data into tidy structures.
}
}
\examples{
\dontrun{
## for instructions on access tokens, see the tokens vignette
vignette("tokens")

## for a quick demo check the rtweet vignette
vignette("rtweet")
}

}

