% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{rtweet_client}
\alias{rtweet_client}
\title{Client}
\usage{
rtweet_client(client_id, client_secret, app, scopes = NULL)
}
\arguments{
\item{client_id, client_secret}{Application OAuth client ID and client Secret.
These are generally not required for \code{rtweet_user()} since the defaults will use
the built-in rtweet app.}

\item{app}{Name of the client, it helps if you make it match with the name of your app.
On the Twitter app the Callback URI must be \verb{http://127.0.0.1:1410/}
(the trailing / must be included).}

\item{scopes}{Default scopes allowed for users using this client.
Leave \code{NULL} to allow everything or choose yours with \code{set_scopes()}.}
}
\description{
Set up your client mechanism for the Twitter API.
}
\examples{
\donttest{
if (interactive()) {
  rtweet_client()
}
}
}
\seealso{
scopes
}
