% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{save_as_csv}
\usage{
save_as_csv(x, file_name)
}
\arguments{
\item{x}{Data table to be saved (tweets or user object)
generated via rtweet function like \code{search_tweets}. If x
is a list object containing both tweets and users data (which
is currently the output for many of the rtweet functions),
then a CSV file is created and saved for each object using the
file_name provided as a base--e.g, if x is a list object from
search_tweets with \code{file_name = "election"}, this
function will save both the tweets data ("election.tweets.csv")
and the user data ("election.users.csv"). If not included in
file_name, the ".csv" extension will be added when writing file
to disk.}

\item{file_name}{Path/file name where object(s) is to be saved.
If object includes both tweets and users data then provided
file_name will be used as base for the two saved files.
For example, \code{file_name = "election"} would save files
as "election.tweets.csv" and "election.users.csv".}
}
\description{
Converts and saves data table generated from rtweet
  package as csv file(s).
}

