% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_tweets}
\alias{search_tweets}
\title{search_tweets}
\usage{
search_tweets(q, n, ...)
}
\arguments{
\item{q}{Search query.}

\item{n}{Number of desired tweets}

\item{...}{For all args see \code{\link{search_tweets.default}}.}
}
\description{
search_tweets
}
