% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_users.R
\name{search_users}
\alias{search_users}
\title{Get users data on accounts identified via search query.}
\usage{
search_users(q, n = 100, parse = TRUE, token = NULL, verbose = TRUE)
}
\arguments{
\item{q}{Query to be searched, used in filtering relevant tweets to
return from Twitter's REST API. Should be a character string not
to exceed 500 characters maximum. Spaces are assumed to function
like boolean "AND" operators. To search for tweets including one
of multiple possible terms, separate search terms with spaces and
the word "OR". For example, the search \code{query = "data science"} searches for tweets using both "data" and
"science" though the words can appear anywhere and in any order
in the tweet. However, when OR is added between search terms,
\code{query = "data OR science"}, Twitter's REST API should
return any tweet that includes either "data" or "science"
appearing in the tweets. At this time, Twitter's users/search API
does not allow complex searches or queries targeting exact
phrases as is allowed by \code{search_tweets}.}

\item{n}{Number of users to return (at most 1,000).}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}

\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{verbose}{If \code{TRUE}, will display a progress bar while downloading.}
}
\value{
Data frame of users returned by query.
}
\description{
Returns data for up to 1,000 users matched by user provided search
query.
}
\examples{

\dontrun{
users <- search_users("#rstats", n = 300)
users

# latest tweet from each user
tweets_data(users)
}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/follow-search-get-users/api-reference/get-users-search}
}
\seealso{
Other users: 
\code{\link{as_screenname}()},
\code{\link{lists_subscribers}()},
\code{\link{lookup_users}()}
}
\concept{users}
