% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwordslangs}
\alias{stopwordslangs}
\title{stopwordslangs data}
\format{A tibble with three variables and 24,000 observations}
\usage{
stopwordslangs
}
\description{
This data comes form a group of Twitter searches conducted on 2017-09-27.
The data are commonly observed words associated with 10 different languages,
including "ar", "en", "es", "fr", "in", "ja", "pt", ru", "tr", and "und".
Variables include word (potential stop words), lang (two or three word code),
and p (p value associated with frequency position along a normal distribution
with higher values meaning the word occurs more frequently and lower values
meaning the words occur less frequently).
}
\examples{
stopwordslangs
}
\keyword{datasets}
