% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwordslangs}
\alias{stopwordslangs}
\title{Twitter stop words in multiple languages data.}
\format{A tibble with three variables and 24,000 observations}
\usage{
stopwordslangs
}
\description{
This data comes from a group of Twitter searches conducted at
several times during the calendar year of 2017. The data are
commonly observed words associated with 10 different languages,
including \code{c("ar", "en", "es", "fr", "in", "ja", "pt", "ru",
"tr", "und")}. Variables include \code{"word"} (potential stop
words), \code{"lang"} (two or three word code), and \code{"p"}
(probability value associated with frequency position along a
normal distribution with higher values meaning the word occurs more
frequently and lower values meaning the words occur less
frequently).
}
\examples{
stopwordslangs
}
\keyword{datasets}
