% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream_tweets2}
\alias{stream_tweets2}
\title{Stream with hardwired reconnection method to ensure timeout integrity.}
\usage{
stream_tweets2(...)
}
\arguments{
\item{...}{Args passed to \code{\link{stream_tweets}} function.}
}
\value{
Returns data as expected using original search_tweets function.
}
\description{
Stream with hardwired reconnection method to ensure timeout integrity.
}
