% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream_tweets2}
\alias{stream_tweets2}
\title{A more robust version of stream_tweets}
\usage{
stream_tweets2(..., dir = NULL, append = FALSE)
}
\arguments{
\item{dir}{Name of directory in which json files should be written.
The default, NULL, will create a timestamped "stream" folder in the
current working directory. If a dir name is provided that does not
already exist, one will be created.}

\item{append}{Logical indicating whether to append or overwrite
file_name if the file already exists. Defaults to FALSE, meaning
this function will overwrite the preexisting file_name (in other
words, it will delete any old file with the same name as
file_name) meaning the data will be added as new lines to file if
pre-existing.}
}
\value{
Returns data as expected using original search_tweets
function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please use \code{\link[=stream_tweets]{stream_tweets()}} instead.
}
\seealso{
\code{\link{rtweet-deprecated}}
}
\keyword{internal}
