% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested.R
\name{all_suggested_users}
\alias{all_suggested_users}
\title{Get all user [account] suggestions for authenticating user}
\usage{
all_suggested_users(parse = TRUE)
}
\arguments{
\item{parse}{Logical indicating whether to parse the returned data into
a tibble data frame. See details for more on the returned users data.}
}
\description{
Returns users data for all users in Twitter's suggested categories.
}
