% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested.R
\name{suggested_users_all}
\alias{suggested_users_all}
\title{Get all user [account] suggestions for authenticating user}
\usage{
suggested_users_all(slugs = NULL, parse = TRUE, token = NULL)
}
\arguments{
\item{slugs}{Optional, one or more slugs returned by
\code{\link{suggested_slugs}}. API rate limits this to 15 max (function
will return warnings for slugs provided beyond the remaining limit).}

\item{parse}{Logical indicating whether to parse the returned data into
a tibble data frame. See details for more on the returned users data.}

\item{token}{Every user should have their own Oauth (Twitter API) token. By
default \code{token = NULL} this function looks for the path to a saved
Twitter token via environment variables (which is what `create_token()`
sets up by default during initial token creation). For instruction on how
to create a Twitter token see the tokens vignette, i.e.,
`vignettes("auth", "rtweet")` or see \code{?tokens}.}
}
\description{
Returns users data for all users in Twitter's suggested categories.
}
