% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggested.R
\name{suggested_slugs}
\alias{suggested_slugs}
\alias{suggested_users}
\alias{suggested_users_all}
\title{Get user [account] suggestions for authenticating user}
\usage{
suggested_slugs(lang = NULL, token = NULL)

suggested_users(slug, lang = NULL, parse = TRUE, token = NULL)

suggested_users_all(slugs = NULL, parse = TRUE, token = NULL)
}
\arguments{
\item{lang}{optional Restricts the suggested categories to the
requested language. The language must be specified by the
appropriate two letter ISO 639-1 representation.}

\item{token}{An OAuth token object. The default, \code{NULL}, will retrieve
a default token with \code{\link{get_token}()}. You should only need to 
use this argument if you are wrapping rtweet in a package. 

See \code{vignette("auth")} for more details.}

\item{slug}{required The short name of list or a category}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame. 
This protects you from the vagaries of the twitter API. Use \code{FALSE} 
to return the "raw" list produced by the JSON returned from the twitter 
API.}

\item{slugs}{Optional, one or more slugs returned by
\code{\link{suggested_slugs}}. API rate limits this to 15 max (function
will return warnings for slugs provided beyond the remaining limit).}
}
\value{
List of recommended categories which can be passed along as
  the "slug" parameter in \code{\link{suggested_users}}

Recommended users
}
\description{
Returns Twitter's list of suggested user categories.

Returns users data for all users in Twitter's suggested categories.
}
\details{
Currently, this parsing process drops all
  recursive (list) columns, which mostly means you are shorted some
  entities data. To maximize users data, however, it is recommended to
  make an additional \code{\link{lookup_users}} call using the user IDs
  returned by this function.
}
\examples{

\dontrun{

## get slugs
slugs <- suggested_slugs()

## use slugs to get suggested users
suggested_users(slugs$slug[1])

## alternatively, get all users from all slugs in one function
sugs <- all_suggested_users()

## print data
sugs

## for complete users data, lookup user IDs
sugs_usr <- lookup_users(sugs$user_id)

## view users data
sugs_usr

}

}
\concept{suggested}
