% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends_available}
\alias{trends_available}
\title{Available Twitter trends along with associated WOEID.}
\usage{
trends_available(token = NULL, parse = TRUE)
}
\arguments{
\item{token}{Expert use only. Use this to override authentication for
a single API call. In most cases you are better off changing the
default for all calls. See \code{\link[=auth_use]{auth_use()}} for details.}

\item{parse}{The default, \code{TRUE}, indicates that the result should
be parsed into a convenient R data structure like a list or data frame.
This protects you from the vagaries of the twitter API. Use \code{FALSE}
to return the "raw" list produced by the JSON returned from the twitter
API.}
}
\value{
Data frame with WOEID column. WOEID is a Yahoo! Where On
Earth ID.
}
\description{
Available Twitter trends along with associated WOEID.
}
\examples{
\dontrun{
## Retrieve available trends
trends <- trends_available()
trends

}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/trends/locations-with-trending-topics/api-reference/get-trends-available}
}
\seealso{
Other trends: 
\code{\link{get_trends}()}
}
\concept{trends}
