% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends_available}
\alias{trends_available}
\title{trends_available}
\usage{
trends_available(token = NULL, parse = TRUE)
}
\arguments{
\item{token}{OAuth token (1.0 or 2.0). By default
\code{token = NULL} fetches a non-exhausted token from
an environment variable tokens.}

\item{parse}{Logical, indicating whether to return parsed
(data.frames) or nested list (fromJSON) object. By default,
\code{parse = TRUE} saves users from the time
[and frustrations] associated with disentangling the Twitter
API return objects.}
}
\value{
Data frame with WOEIDs. WOEID is a Yahoo! Where On
  Earth ID.
}
\description{
Returns Twitter trends based on requested WOEID.
}

