% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends_available}
\alias{trends_available}
\title{Available Twitter trends along with associated WOEID.}
\usage{
trends_available(token = NULL, parse = TRUE)
}
\arguments{
\item{token}{Use this to override authentication for
a single API call. In many cases you are better off changing the
default for all calls. See \code{\link[=auth_as]{auth_as()}} for details.}

\item{parse}{If \code{TRUE}, the default, returns a tidy data frame. Use \code{FALSE}
to return the "raw" list corresponding to the JSON returned from the
Twitter API.}
}
\value{
Data frame with WOEID column. WOEID is a Yahoo! Where On
Earth ID.
}
\description{
Available Twitter trends along with associated WOEID.
}
\examples{
if (auth_has_default()) {
## Retrieve available trends
trends <- trends_available()
trends

}

}
\references{
\url{https://developer.twitter.com/en/docs/twitter-api/v1/trends/locations-with-trending-topics/api-reference/get-trends-available}
}
\seealso{
Other trends: 
\code{\link{get_trends}()}
}
\concept{trends}
