% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends_available}
\alias{trends_available}
\title{Available Twitter trends along with associated WOEID.}
\usage{
trends_available(token = NULL, parse = TRUE)
}
\arguments{
\item{token}{OAuth token. By default \code{token = NULL} fetches a
non-exhausted token from an environment variable. Find instructions
on how to create tokens and setup an environment variable in the
tokens vignette (in r, send \code{?tokens} to console).}

\item{parse}{Logical, indicating whether to return parsed
(data.frames) or nested list object. By default,
\code{parse = TRUE} saves users from the time
[and frustrations] associated with disentangling the Twitter
API return objects.}
}
\value{
Data frame with WOEID column. WOEID is a Yahoo! Where On
  Earth ID.
}
\description{
Available Twitter trends along with associated WOEID.
}
\examples{
\dontrun{
## Retrieve available trends
trends <- trends_available()
trends

}

}
\seealso{
Other trends: 
\code{\link{get_trends}()}
}
\concept{trends}
