% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ts_plot}
\alias{ts_plot}
\title{ts_plot}
\usage{
ts_plot(rt, by = "days", dtname = "created_at", txt = "text",
  filter = NULL, key = NULL, lwd = 1.5, cols = NULL, theme = "light",
  main = NULL, xlab = "Time", ylab = "Number of Tweets", box = TRUE,
  axes = TRUE, legend.title = NULL, ticks = 0, cex = 0.9,
  cex.main = 1.25, cex.sub = 0.9, cex.lab = 0.9, cex.axis = 0.7,
  mai = c(0.515, 0.575, 0.15, 0.3), plot = TRUE, ...)
}
\arguments{
\item{rt}{Tweets or users data frame}

\item{by}{Unit of time, e.g., \code{secs, days, weeks,
months, years}}

\item{dtname}{Name of date-time (POSIXt) object. Defaults
to created_at.}

\item{txt}{Name of text (chr) variable in data frame which
filter is applied to. Defaults to text.}

\item{filter}{Vector of regular expressions with which to
filter data (creating multiple time series).}

\item{key}{Optional provide pretty labels for filters.
Defaults to actual filters.}

\item{lwd}{Width of time series line(s). Defaults to 1.5}

\item{cols}{Colors for filters. Leave NULl for default color
scheme.}

\item{theme}{Character string specifyng whether and which
plot theme should be used; options include "lighter",
"darker", and "nerdy"}

\item{main}{Optional, text for plot title.}

\item{xlab}{Optional, text for x-axis title, defaults to
"Time".}

\item{ylab}{Optional, text for y-axis title, defaults to
"Number of Tweets"}

\item{box}{Logical indicating whether to draw box around
plot area. Defaults to true.}

\item{axes}{Logical indicating whether to draw axes. Defaults
to true.}

\item{legend.title}{Provide title for legend ro ignore to
leave blank (default)}

\item{ticks}{Numeric specifying width of tick marks. Defaults
to zero. If you'd like tick marks, try setting this value
to 1.25.}

\item{cex}{Global cex setting defaults to .90}

\item{cex.main}{Size of plot title (if plot title provided
via \code{main = "title"} argument).}

\item{cex.sub}{Size of subtitles}

\item{cex.lab}{Size of axis labels and legend text}

\item{cex.axis}{Size of other axis text.}

\item{mai}{Margins in inches.}

\item{plot}{Logical indicating whether to draw plot.}

\item{\dots}{Arguments passed to plot (and par) function}
}
\description{
Plots frequency of tweets as time series or, if multiple
  filters (text-based criteria used to subset data) are
  specified, multiple time series.
}
\examples{
\dontrun{
## stream tweets mentioning beibs and tswift for 10 mins
rt <- rtweet::stream_tweets(
    q = "justinbieber,taylorswift13",
    timeout = (60 * 60 * 10))

## split mentions into distinct elements
mentions <- strsplit(rt$mentions_screen_name, ",")

## sorted table of mentions
mentions <- sort(table(unlist(mentions)),
    decreasing = TRUE)

## exclude biebs and tswift
mentions <- grep("justinbieber|taylorswift13", names(mentions),
    invert = TRUE, value = TRUE)

## store next most pop in obj
thirdpop <- mentions[1]

##plot with mentions as filters
ts.df <- ts_plot(rt, by = "mins", filter = c(
    "justinbieber", "taylorswift", thirdpop),
    main = "Biebs vs Tswift")

## preview returned data frame
head(ts.df)

}
}

