% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ts_plot}
\alias{ts_plot}
\title{ts plot}
\usage{
ts_plot(data, by = "days", group = NULL, ...)
}
\arguments{
\item{data}{Data frame or grouped data frame.}

\item{by}{Desired interval of time expressed as numeral plus secs,
mins, hours, days, weeks, months, years. If a numeric is
provided, the value is assumed to be in seconds.}

\item{group}{Name of grouping variable for use in plotting multiple
time series. This is automatically configured if data is a
grouped_df (dplyr's group_by function).}

\item{...}{Other args passed to ggplot geom_line or, if ggplot2 is
not installed, then to lines plotting function.}
}
\value{
Either a ggplot plot object or an invisibly returned time
  series-like data frame.
}
\description{
Returns time series-like data frame.
}
\examples{
\dontrun{
rt <- search_tweets("rstats", n = 10000)
ts_plot(rt, "10 mins")

rt \%>\%
  dplyr::group_by(is_retweet) \%>\%
  ts_plot("hours")

## compare account activity for some important national political figures
tmls <- get_timeline(c("SenSchumer", "SenGillibrand", "realDonaldTrump"), n = 3000)
## examine all twitter activity
ts_plot(tmls, "months")
## group by screen name
ts_plot(dplyr::group_by(tmls, screen_name), "months") + theme_mwk()
## group by screen name and is_retweet
ts_plot(dplyr::group_by(tmls, screen_name, is_retweet), "months") + theme_minimal()
}
}
