% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ts_plot}
\alias{ts_plot}
\title{ts_plot}
\usage{
ts_plot(rt, by = "days", txt = "text", filter = NULL, exclude = NULL,
  key = NULL, ...)
}
\arguments{
\item{rt}{Tweets data frame}

\item{by}{Unit of time, e.g., \code{secs, days, weeks,
months, years}}

\item{txt}{Name of text variable in data frame which
filter is applied to.}

\item{filter}{Vector of regular expressions with which to
filter data (creating multiple time series)}

\item{exclude}{Vector of regular expressions with which to
distinguish data.}

\item{key}{Labels for filters. Defaults to actual filters.}

\item{\dots}{Other arguments passed to plot function.}
}
\description{
ts_plot
}
\examples{
\dontrun{
# stream tweets mentioning presidential debates hashtag
#   for 5 minutes
x <- stream_tweets(q = "debates2016", timeout = (60 * 5))
tsplot(x)
}
}

