% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ts_plot}
\alias{ts_plot}
\title{ts_plot}
\usage{
ts_plot(rt, by = "days", dtname = "created_at", txt = "text",
  filter = NULL, key = NULL, cols = NULL, box = TRUE, axes = TRUE,
  main = NULL, mai = c(0.475, 0.5, 0.15, 0.3), cex = 0.9,
  cex.axis = 0.8, cex.lab = 0.95, cex.main = 1.5, cex.sub = 1,
  lwd = 1.5, legend.title = NULL, ...)
}
\arguments{
\item{rt}{Tweets or users data frame}

\item{by}{Unit of time, e.g., \code{secs, days, weeks,
months, years}}

\item{dtname}{Name of date-time (POSIXt) object. Defaults
to created_at.}

\item{txt}{Name of text (chr) variable in data frame which
filter is applied to. Defaults to text.}

\item{filter}{Vector of regular expressions with which to
filter data (creating multiple time series).}

\item{key}{Optional provide pretty labels for filters.
Defaults to actual filters.}

\item{cols}{Colors for filters. Leave NULl for default color
scheme.}

\item{box}{Logical indicating whether to draw box around
plot area. Defaults to true.}

\item{axes}{Logical indicating whether to draw axes. Defaults
to true.}

\item{main}{Optional argument to provide plot title.}

\item{mai}{Margins in inches.}

\item{cex}{Global cex setting defaults to .90}

\item{cex.axis}{Size of other axis text.}

\item{cex.lab}{Size of axis labels and legend text}

\item{cex.main}{Size of plot title (if plot title provided
via \code{main = "title"} argument).}

\item{cex.sub}{Size of subtitles}

\item{lwd}{Width of time series line(s). Defaults to 1.5}

\item{legend.title}{Provide title for legend ro ignore to
leave blank (default)}

\item{legend}{Logical indicating wether to include a plot
legend. Defaults to true.}

\item{plot}{Logical indicating whether to draw plot.}

\item{\dots}{Arguments passed to plot function, e.g.,
\code{main = "#rstats tweets"},
\code{xlab = "Date"},
\code{ylab = "Tweets"},
\code{main.cex = 1}.}
}
\description{
Plots frequency of tweets as time series or, if multiple
  filters (text-based criteria used to subset data) are
  specified, multiple time series.
}
\examples{
\dontrun{
## stream tweets mentioning beibs and tswift for 10 mins
rt <- rtweet::stream_tweets(
    q = "justinbieber,taylorswift13",
    timeout = (60 * 60 * 10))

## split mentions into distinct elements
mentions <- strsplit(rt$mentions_screen_name, ",")

## sorted table of mentions
mentions <- sort(table(unlist(mentions)),
    decreasing = TRUE)

## exclude biebs and tswift
mentions <- grep("justinbieber|taylorswift13", names(mentions),
    invert = TRUE, value = TRUE)

## store next most pop in obj
thirdpop <- mentions[1]

##plot with mentions as filters
ts.df <- ts_plot(rt, by = "mins", filter = c(
    "justinbieber", "taylorswift", thirdpop),
    main = "Biebs vs Tswift")

## preview returned data frame
head(ts.df)

}
}

