% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tweets_data}
\alias{tweets_data}
\title{tweets_data}
\usage{
tweets_data(x)
}
\arguments{
\item{x}{Data frame from \code{lookup_users} or
\code{get_tweets}.}
}
\value{
Tweets data tibble
}
\description{
Returns tweets data from rtweet function output.
}

