% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{tweets_data}
\alias{tweets_data}
\alias{tweet_data}
\alias{data_tweet}
\alias{data_tweets}
\title{tweets_data}
\usage{
tweets_data(users)
}
\arguments{
\item{users}{Data frame of status data generated via
\code{search_tweets}, \code{stream_tweets}, \code{get_timeline},
\code{lookup_statuses}, etc.}
}
\value{
Tweets data frame.
}
\description{
Tweets data frame from users returned in a users data object.
  Typically, this involves the most recent tweet of each user, though
  in some cases the most recent tweet may not be available.
}
\examples{
\dontrun{
## get twitter user data
jack <- lookup_users("jack")

## get data on most recent tweet from user(s)
tweets_data(jack)

## search for 100 tweets containing the letter r
r <- search_tweets("r")

## print tweets data (only first 10 rows are shown)
head(r, 10)

## preview users data
head(users_data(r))
}

}
\seealso{
Other tweets: \code{\link{get_favorites.default}},
  \code{\link{get_timeline_}},
  \code{\link{lookup_statuses}},
  \code{\link{search_tweets.default}},
  \code{\link{stream_tweets}}
}
