% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets_and_users.R
\name{tweets_with_users}
\alias{tweets_with_users}
\alias{users_with_tweets}
\title{Parsing data into tweets/users data tibbles}
\usage{
tweets_with_users(x)

users_with_tweets(x)
}
\arguments{
\item{x}{A list of responses, with one element for each page.}
}
\value{
A tweets/users tibble with users/tweets attribute.
}
\description{
For internal use only
}
