% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.users.R
\name{user_lookup}
\alias{user_lookup}
\title{user_lookup}
\usage{
user_lookup(user_ids, token = NULL)
}
\arguments{
\item{token}{OAuth tokens (1.0 or 2.0)}

\item{users}{Screen name or user id of target users.}

\item{df}{logical, indicating whether to format response
as data frame}
}
\value{
response object
}
\description{
user_lookup
}
\seealso{
See \url{https://dev.twitter.com/overview/
documentation} for more
  information on using Twitter's API.
}

