% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{users_data}
\alias{users_data}
\title{users_data}
\usage{
users_data(x)
}
\arguments{
\item{x}{Data frame from \link{\code{search_tweets}} or
\link{\code{stream_tweets}}.}
}
\value{
Users data tibble
}
\description{
Returns users data from rtweet function output.
}

