% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{users_data}
\alias{users_data}
\alias{user_data}
\alias{data_user}
\alias{data_users}
\title{users_data}
\usage{
users_data(tweets)
}
\arguments{
\item{tweets}{Data frame of Twitter statuses (tweets) generated via
\code{\link{get_timeline}}, \code{\link{search_tweets}}, or
\code{\link{stream_tweets}}.}
}
\value{
Users data frame from tweets returned in a tweets data object.
}
\description{
Returns users data frame from returned tweets
  data object.
}
\examples{
\dontrun{
## search for 100 tweets containing the letter r
r <- search_tweets("r")

## print tweets data (only first 10 rows are shown)
head(r, 10)

## extract users data
head(users_data(r))
}

}
\seealso{
Other users: \code{\link{lists_subscribers}},
  \code{\link{lookup_users}}, \code{\link{search_users}}

Other extractors: \code{\link{max_id}},
  \code{\link{tweets_data}}
}
