% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{users_data}
\alias{data_user}
\alias{data_users}
\alias{user_data}
\alias{users_data}
\title{users_data}
\usage{
users_data(tweets)
}
\arguments{
\item{tweets}{Data frame of Twitter statuses (tweets) generated via
\code{\link{get_timeline}}, \code{\link{search_tweets}}, or
\code{\link{stream_tweets}}.}
}
\value{
Users data frame from tweets returned in a tweets data object.
}
\description{
Returns users data frame from returned tweets
  data object.
}
\examples{
\dontrun{
# search for 100 tweets containing the letter r
r <- search_tweets("r")

# print tweets data (only first 10 rows are shown)
r

# extract users data
users_data(r)
}

}
\seealso{
Other users: \code{\link{lookup_users}},
  \code{\link{search_users}}
}

